% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikmisc_DATA.R
\docType{data}
\name{dat_ergoStool_e}
\alias{dat_ergoStool_e}
\title{nlme::ergoStool, labeled}
\format{
tibble
}
\source{
\code{?nlme::ergoStool}
}
\usage{
data(dat_ergoStool_e)
}
\description{
Using \code{nlme::ergoStool}, then applying labeling variables with \code{labelled::var_label} based on \code{?nlme::ergoStool}
}
\examples{
data(dat_ergoStool_e)
str(dat_ergoStool_e)
}
\keyword{datasets}
