% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikmisc_DATA.R
\docType{data}
\name{dat_mtcars_e}
\alias{dat_mtcars_e}
\title{datasets::mtcars, factored and labeled}
\format{
tibble
}
\source{
\code{?datasets::mtcars}
}
\usage{
data(dat_mtcars_e)
}
\description{
Using \code{datasets::mtcars}, then applying labeled factors and labeling variables with \code{labelled::var_label} based on \code{?datasets::mtcars}
}
\examples{
data(dat_mtcars_e)
str(dat_mtcars_e)
}
\keyword{datasets}
