% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikmisc_DATA.R
\docType{data}
\name{dat_unicode_tables_sel_byhand_e}
\alias{dat_unicode_tables_sel_byhand_e}
\title{Unicode characters from most of the tables from Wikipedia's "List of Unicode characters"}
\format{
tibble
}
\source{
\code{https://en.wikipedia.org/wiki/List_of_Unicode_characters} from 4/25/2021
}
\usage{
data(dat_unicode_tables_sel_byhand_e)
}
\description{
The table includes Code, Glyph (symbol), Description of the character, CharacterSet group, whether it is_unicode (always TRUE), the CodeEscaped as stringi needs it, and the hand-selected ASCII replacement.
}
\examples{
data(dat_unicode_tables_sel_byhand_e)
str(dat_unicode_tables_sel_byhand_e)
}
\keyword{datasets}
