% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_bestsubset.R
\name{e_bestsubset}
\alias{e_bestsubset}
\title{Best subset selection returns results sorted by BIC}
\usage{
e_bestsubset(form, dat, nbest = 5)
}
\arguments{
\item{form}{lm formula for full model}

\item{dat}{data to use}

\item{nbest}{number of models to return for each model size, default is 5}
}
\value{

}
\description{
Best subset selection returns results sorted by BIC
}
\examples{
i_best <-
  e_bestsubset(
    form = formula(mpg ~ cyl + disp + hp + gear)
  , dat  = datasets::mtcars
  )

  op <- options(); # saving old options
  options(width=100) # setting command window output text width wider
i_best \%>\% print(n = Inf, width = Inf)
  options(op); # reset (all) initial options
}
