% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_ALL.R
\name{e_calc_a1c_categories}
\alias{e_calc_a1c_categories}
\title{A1c category labels using CDC classification or user-defined categories}
\usage{
e_calc_a1c_categories(
  a1c,
  a1c_breaks = c(-Inf, 5.7, 6.4, Inf),
  a1c_labels = c("Neither", "Pre-diabetes", "Diabetes")
)
}
\arguments{
\item{a1c}{A1c values}

\item{a1c_breaks}{Breaks defining labeled ranges, intervals are closed on the left}

\item{a1c_labels}{Labels for values between break values including lower bound}
}
\value{
a1c_cat, categorical levels of A1c
}
\description{
See CDC \code{https://www.cdc.gov/diabetes/managing/managing-blood-sugar/a1c.html}
}
\examples{
a1c      <- c(NA, seq(0,15, by =1), NA)
a1c_cat  <- e_calc_a1c_categories(a1c)
data.frame(a1c, a1c_cat)
}
