% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_ALL.R
\name{e_calc_bmi_categories}
\alias{e_calc_bmi_categories}
\title{BMI category labels using CDC classification}
\usage{
e_calc_bmi_categories(
  bmi,
  bmi_breaks = c(-Inf, 18.5, 25, 30, Inf),
  bmi_labels = c("Underweight", "Normal or Healthy Weight", "Overweight", "Obese")
)
}
\arguments{
\item{bmi}{BMI value}

\item{bmi_breaks}{Breaks defining labeled ranges, intervals are closed on the left}

\item{bmi_labels}{Labels for values between break values including lower bound}
}
\value{
bmi_cat, categorical levels of BMI
}
\description{
BMI category labels using CDC classification
}
\examples{
e_calc_bmi(60, 150) \%>\% e_calc_bmi_categories()
e_calc_bmi(seq(40, 70, by = 5), 150) \%>\% e_calc_bmi_categories()
}
