% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_calc_ECURE_LUDTM.R
\name{e_calc_ECURE_scale_LUDTM}
\alias{e_calc_ECURE_scale_LUDTM}
\title{ECURE scale the LUDTM when less than expected}
\usage{
e_calc_ECURE_scale_LUDTM(UD = NA, UDE = NA, T = NA, C_min = 45, C_max = 120)
}
\arguments{
\item{UD}{number of upper division credits}

\item{UDE}{upper division expectation for credits}

\item{T}{total credits earned}

\item{C_min}{start of UD (upper division) credits}

\item{C_max}{credits required for graduation}
}
\value{
LUDTM  Lower- to upper-division transition metric when less than expected
}
\description{
See \code{e_calc_ECURE_LUDTM}
}
