% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_calc_ECURE_LUDTM.R
\name{e_calc_ECURE_UDE}
\alias{e_calc_ECURE_UDE}
\title{ECURE Upper Division Expectation (UDE)}
\usage{
e_calc_ECURE_UDE(T = NA, C_min = 45, C_max = 120, UDC_min = 54)
}
\arguments{
\item{T}{total credits earned}

\item{C_min}{start of UD (upper division) credits}

\item{C_max}{credits required for graduation}

\item{UDC_min}{minimum UD credits at graduation}
}
\value{
UDE    Upper Division Expectation
}
\description{
See \code{e_calc_ECURE_LUDTM}
}
\examples{
# Starting your senior semester, expected number of upper-division credits
e_calc_ECURE_UDE(T = 90)
}
