% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_ALL.R
\name{e_calc_phq2_categories}
\alias{e_calc_phq2_categories}
\title{PHQ-2 category labels using Table 4 from Kroenke K, Spitzer RL, Psychiatric Annals 2002;32:509-521}
\usage{
e_calc_phq2_categories(
  phq2,
  phq2_breaks = c(-Inf, 3, Inf),
  phq2_labels = c("None-minimal", "Depressed")
)
}
\arguments{
\item{phq2}{PHQ-2 scores (0 - 6)}

\item{phq2_breaks}{Breaks defining labeled ranges, intervals are closed on the left (typically, 3 or greater indicates major depressive disorder is likely)}

\item{phq2_labels}{Labels for values between break values including lower bound}
}
\value{
phq2_cat, categorical levels of PHQ-2
}
\description{
PHQ-2 Scores and Proposed Treatment Actions.
https://www.pcpcc.org/sites/default/files/resources/instructions.pdf
}
\details{
The PHQ-2 inquires about the frequency of depressed mood and anhedonia over the past two weeks. The PHQ-2 includes the first two items of the PHQ-9.
The purpose of the PHQ-2 is to screen for depression in a "first-step" approach.
Patients who screen positive should be further evaluated with the PHQ-9 to determine whether they meet criteria for a depressive disorder.
PHQ-2 score obtained by adding score for each question (total points)

Interpretation:
A PHQ-2 score ranges from 0-6. The authors identified a score of 3 as the optimal cutpoint when using the PHQ-2 to screen for depression.
If the score is 3 or greater, major depressive disorder is likely.
Patients who screen positive should be further evaluated with the PHQ-9, other diagnostic instruments, or direct interview to determine whether they meet criteria for a depressive disorder.
}
\examples{
e_calc_phq2_categories(c(0, 2, 3, 5))

dplyr::tibble(
  phq2 = seq(0, 6, by = 1)
, phq2_cat = phq2 |> e_calc_phq2_categories()
) |>
print(n = Inf)
}
