% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_ALL.R
\name{e_calc_phq9_categories}
\alias{e_calc_phq9_categories}
\title{PHQ-9 category labels using Table 4 from Kroenke K, Spitzer RL, Psychiatric Annals 2002;32:509-521}
\usage{
e_calc_phq9_categories(
  phq9,
  phq9_breaks = c(-Inf, 5, 10, 15, 20, Inf),
  phq9_labels = c("None-minimal", "Mild", "Moderate", "Moderately Severe", "Severe")
)
}
\arguments{
\item{phq9}{PHQ-9 scores (0 - 27)}

\item{phq9_breaks}{Breaks defining labeled ranges, intervals are closed on the left}

\item{phq9_labels}{Labels for values between break values including lower bound}
}
\value{
phq9_cat, categorical levels of PHQ-9
}
\description{
PHQ-9 Scores and Proposed Treatment Actions.
https://www.pcpcc.org/sites/default/files/resources/instructions.pdf
}
\examples{
e_calc_phq9_categories(c(0, 6, 12, 21))
dplyr::tibble(phq9 = seq(0, 27, by = 1), phq9_cat = phq9 \%>\% e_calc_phq9_categories()) \%>\% print(n = Inf)
}
