% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_calc_rmse.R
\name{e_calc_rmse}
\alias{e_calc_rmse}
\title{Root Mean Squared Error}
\usage{
e_calc_rmse(obs, pred)
}
\arguments{
\item{obs}{list of observed values}

\item{pred}{list of predicted values}
}
\value{
rmse
}
\description{
Root Mean Squared Error
}
\examples{
e_calc_rmse(obs = 1:5 + rnorm(n=5), pred = 1:5)
}
