% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_CI_limits.R
\name{e_CI_limits}
\alias{e_CI_limits}
\title{Extract confidence interval limits from vector (of bootstrap resamples)}
\usage{
e_CI_limits(
  x,
  conf_level = 0.95,
  percentile_lower = NULL,
  percentile_upper = NULL,
  sw_return_exact_CI_coverage = FALSE
)
}
\arguments{
\item{x}{vector of values}

\item{conf_level}{central proportion coverage of interval}

\item{percentile_lower}{\code{NULL}, or lower percentile (0.025) of interval if \code{conf_level} not used}

\item{percentile_upper}{\code{NULL}, or upper percentile (0.975) of interval if \code{conf_level} not used}

\item{sw_return_exact_CI_coverage}{T/F, return exact coverage of interval or not}
}
\value{
out a named vector of 2 (lower, upper) or 3 (and coverage) values
}
\description{
Extract confidence interval limits from vector (of bootstrap resamples)
}
\examples{
x = seq(0, 1, length = 100001) |> sample(replace = TRUE)
e_CI_limits(x)
e_CI_limits(x, sw_return_exact_CI_coverage = TRUE)
e_CI_limits(x, conf_level = 0.80, sw_return_exact_CI_coverage = TRUE)
e_CI_limits(x, percentile_lower = 0.01, percentile_upper = 0.90, sw_return_exact_CI_coverage = TRUE)
}
