% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_complete_multiple_keys_ALL.R
\name{e_coalesce_by_column}
\alias{e_coalesce_by_column}
\title{Combine rows with different data values}
\usage{
e_coalesce_by_column(dat)
}
\arguments{
\item{dat}{data.frame of keys (or other columns) to combine}
}
\value{
coalesced data.frame
}
\description{
For matching \code{group_by()} keys, keeps first non-NA value for each column.
}
\examples{
\dontrun{
dat \%>\%
  dplyr::group_by(A) \%>\%
  dplyr::summarise_all( e_coalesce_by_column )
}
}
