% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_complete_multiple_keys_ALL.R
\name{e_coalesce_column_set}
\alias{e_coalesce_column_set}
\title{Coalesce column set}
\usage{
e_coalesce_column_set(dat)
}
\arguments{
\item{dat}{data.frame keys to combine}
}
\value{
data.frame with most complete key rows
}
\description{
For a set of key variables that are unique for participants,
  identify sets of rows with unique key variables in common
  and collapse them together to have a single "most complete" row
  of key values by removing the NAs.
Each key variable must have a unique value for each participant.
}
