% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_complete_multiple_keys_ALL.R
\name{e_complete_multiple_keys}
\alias{e_complete_multiple_keys}
\title{Complete multiple key ID variables}
\usage{
e_complete_multiple_keys(
  dat_data,
  dat_keys = NULL,
  col_keys = c("a", "b", "c")
)
}
\arguments{
\item{dat_data}{data frame with a set of variables that are keys}

\item{dat_keys}{optional data frame with only key variables, more complete than set in dat_data}

\item{col_keys}{key columns names}
}
\value{
dat_data with updated key columns
}
\description{
When data are combined from multiple sources and multiple identifier keys (ID)
are used, sometimes in combinations, then we may want to provide the
most complete set of the IDs for each observation.
By definition, each key value is unique within its column to identify an observational unit.
}
\examples{

dat_data <-
  tibble::tribble(
    ~a, ~b, ~c, ~x, ~d1, ~d2
  ,  1, 11, NA, NA,   1,  20
  ,  1, NA, 10, NA,   2,  21
  ,  2, NA, NA, NA,   3,  22
  , NA, 22, 20, NA,   4,  23
  ,  2, NA, 20, NA,   5,  24
  ,  3, 33, NA, NA,   6,  25
  ,  4, NA, 40, NA,   7,  26
  ,  5, NA, NA, NA,   8,  27
  ,  6, NA, 60, NA,   9,  28
  ,  6, NA, 60, NA,  10,  29
  , NA, 77, NA, NA,  11,  30
  , NA, 88, 80, NA,  12,  31
  , NA, 88, NA, NA,  13,  32
  , NA, NA, NA, NA,  14,  33
  )

dat_data_updated <-
  e_complete_multiple_keys(
    dat_data
  , dat_keys = NULL
  , col_keys = c("a", "b", "c", "x")
  )

dat_data_updated \%>\% print(n=Inf)

}
