% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_convert_units_ALL.R
\name{e_convert_cm_to_in}
\alias{e_convert_cm_to_in}
\title{Convert centimeters (cm) to inches (in)}
\usage{
e_convert_cm_to_in(cm)
}
\arguments{
\item{cm}{list of values in centimeters (cm)}
}
\value{
inches  list of values in inches (in)
}
\description{
Convert centimeters (cm) to inches (in)
}
\examples{
e_convert_cm_to_in(2.54)
}
