% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_convert_units_ALL.R
\name{e_convert_Grade_to_GPA}
\alias{e_convert_Grade_to_GPA}
\title{convert letter grades to GPA equivalent}
\usage{
e_convert_Grade_to_GPA(Grade)
}
\arguments{
\item{Grade}{character string letter grades with trailing + or -: "F", "D-", "D", "D+", ..., "A-", "A", "A+"}
}
\value{
GPA  numeric from F = 0 to A+ = 4.33, NA if not a letter grade
}
\description{
Using UNM standard: \url{https://unm-student.custhelp.com/app/answers/detail/a_id/3461}
}
\examples{
e_convert_Grade_to_GPA(
    Grade =
      c("A+", "A", "A-"
      , "B+", "B", "B-"
      , "C+", "C", "C-"
      , "D+", "D", "D-"
      , "F"
      , NA, 0, "hello"
      )
  )
}
