% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_convert_units_ALL.R
\name{e_convert_in_to_cm}
\alias{e_convert_in_to_cm}
\title{Convert inches (in) to centimeters (cm)}
\usage{
e_convert_in_to_cm(inches)
}
\arguments{
\item{inches}{list of values in inches (in)}
}
\value{
cm      list of values in centimeters (cm)
}
\description{
Convert inches (in) to centimeters (cm)
}
\examples{
e_convert_in_to_cm(1)
}
