% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_convert_units_ALL.R
\name{e_convert_kg_to_lb}
\alias{e_convert_kg_to_lb}
\title{Convert kilograms (kg) to pounds (lb)}
\usage{
e_convert_kg_to_lb(kg)
}
\arguments{
\item{kg}{list of values in kilograms (kg)}
}
\value{
lb  list of values in pounds (lb)
}
\description{
Convert kilograms (kg) to pounds (lb)
}
\examples{
e_convert_kg_to_lb(0.45359237)
}
