% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_convert_units_ALL.R
\name{e_convert_lb_to_kg}
\alias{e_convert_lb_to_kg}
\title{Convert pounds (lb) to kilograms (kg)}
\usage{
e_convert_lb_to_kg(lb)
}
\arguments{
\item{lb}{list of values in pounds (lb)}
}
\value{
kg  list of values in kilograms (kg)
}
\description{
Convert pounds (lb) to kilograms (kg)
}
\examples{
e_convert_lb_to_kg(1)
}
