% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_convert_oddsratio-probability_ALL.R
\name{e_convert_logistic}
\alias{e_convert_logistic}
\title{Logistic: Convert from log(Odds Ratio) to Probability}
\usage{
e_convert_logistic(z = NULL, sw_percent = FALSE)
}
\arguments{
\item{z}{list of log(odds ratios) logit-scale values}

\item{sw_percent}{T/F output is percent}
}
\value{
p         list of probabilities
}
\description{
Logistic: Convert from log(Odds Ratio) to Probability
}
\examples{
z <- qnorm(seq(0, 1, by = 0.1))
e_convert_logistic(z)
}
