% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_convert_oddsratio_to_probability.R
\name{e_convert_oddsratio_to_probability}
\alias{e_convert_oddsratio_to_probability}
\title{Convert from Odds Ratio to Probability}
\usage{
e_convert_oddsratio_to_probability(or = NULL, sw_percent = FALSE)
}
\arguments{
\item{or}{list of odds ratios}

\item{sw_percent}{T/F return as percent, instead?}
}
\value{
p         list of probabilities
}
\description{
Convert from Odds Ratio to Probability
}
\examples{
or <- c(0, 1/9, 1/4, 1/2, 1, 2, 4, 9)
e_convert_oddsratio_to_probability(or)
}
