% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_convert_probability_to_oddsratio.R
\name{e_convert_probability_to_oddsratio}
\alias{e_convert_probability_to_oddsratio}
\title{Convert from Probability to Odds Ratio}
\usage{
e_convert_probability_to_oddsratio(p = NULL, sw_percent = FALSE)
}
\arguments{
\item{p}{list of probabilities}

\item{sw_percent}{T/F input is percent}
}
\value{
or        list of odds ratios
}
\description{
Convert from Probability to Odds Ratio
}
\examples{
p <- c(0, 0.1, 0.2, 1/3, 1/2, 2/3, 0.8, 0.9)
e_convert_probability_to_oddsratio(p)
}
