% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_data_class_align_between_datasets.R
\name{e_data_class_align_between_datasets}
\alias{e_data_class_align_between_datasets}
\title{For two datasets with some matching variable names, align classes before row binding together}
\usage{
e_data_class_align_between_datasets(
  dat1 = NULL,
  dat2 = NULL,
  sw_bind_rows = FALSE
)
}
\arguments{
\item{dat1}{data.frame or tibble, treated as primary dataset for class}

\item{dat2}{data.frame or tibble, treated as secondary dataset for class}

\item{sw_bind_rows}{T/F whether to return \code{dplyr::bind_rows(dat1, dat2)} after class alignment}
}
\value{
out        list with \code{dat1} and \code{dat2}, and \code{dat_class} at the start, differences, and end of process
}
\description{
For two datasets with some matching variable names, align classes before row binding together
}
\details{
Logic:
\enumerate{
\item Determine common variable names between \code{dat1} and \code{dat2}.
\item Determine classes of common variables, retain list of different classes or factor variables with different levels.
\item If both factors and different levels, expand list of factors to include those in order from \code{dat1} then \code{dat2}.
\item If one dataset has all NA, then it is set to the class of the other dataset. If both all NA, then set to dat1 class.
\item Otherwise, the more specific class is converted to the more general class:
logical < integer < numeric < complex < factor (labels) < character.
}
}
\examples{
# Create list of variable names for all combinations of data classes and factor levels
#lab <- c("L", "I", "N", "X", "F", "f", "C", "n") # labels for datatypes; X = complex, F and f have different factor levels, n = all NAs
#lab_pairs <- expand.grid(lab, lab)
#paste0("Var_", lab_pairs$Var1, "_", lab_pairs$Var2)

n_obs <- 100

# Variables are named as "Var1_[dat1 class]_[dat2 class]" for all class comparisons.
dat1 <-
  tibble::tibble(
    Var_d1  = rep("dat1 only", n_obs)
  , Var_L_L = rbinom(n_obs, size = 1, prob = 0.5) |> as.logical()
  , Var_I_L = rbinom(n_obs, size = 1, prob = 0.5) |> as.integer()
  , Var_N_L = rbinom(n_obs, size = 1, prob = 0.5) |> as.numeric()
  , Var_X_L = complex(real = stats::rnorm(n_obs), imaginary = stats::rnorm(n_obs))
  , Var_F_L = sample(letters[ 1: 6], n_obs, replace = TRUE) |> factor(levels = letters[ 1: 6])
  , Var_f_L = sample(letters[21:26], n_obs, replace = TRUE) |> factor(levels = letters[21:26])
  , Var_C_L = sample(letters[11:16], n_obs, replace = TRUE)
  , Var_n_L = rep(NA, n_obs)
  , Var_L_I = Var_L_L, Var_I_I = Var_I_L, Var_N_I = Var_N_L, Var_X_I = Var_X_L, Var_F_I = Var_F_L, Var_f_I = Var_f_L, Var_C_I = Var_C_L, Var_n_I = Var_n_L
  , Var_L_N = Var_L_L, Var_I_N = Var_I_L, Var_N_N = Var_N_L, Var_X_N = Var_X_L, Var_F_N = Var_F_L, Var_f_N = Var_f_L, Var_C_N = Var_C_L, Var_n_N = Var_n_L
  , Var_L_X = Var_L_L, Var_I_X = Var_I_L, Var_N_X = Var_N_L, Var_X_X = Var_X_L, Var_F_X = Var_F_L, Var_f_X = Var_f_L, Var_C_X = Var_C_L, Var_n_X = Var_n_L
  , Var_L_F = Var_L_L, Var_I_F = Var_I_L, Var_N_F = Var_N_L, Var_X_F = Var_X_L, Var_F_F = Var_F_L, Var_f_F = Var_f_L, Var_C_F = Var_C_L, Var_n_F = Var_n_L
  , Var_L_f = Var_L_L, Var_I_f = Var_I_L, Var_N_f = Var_N_L, Var_X_f = Var_X_L, Var_F_f = Var_F_L, Var_f_f = Var_f_L, Var_C_f = Var_C_L, Var_n_f = Var_n_L
  , Var_L_C = Var_L_L, Var_I_C = Var_I_L, Var_N_C = Var_N_L, Var_X_C = Var_X_L, Var_F_C = Var_F_L, Var_f_C = Var_f_L, Var_C_C = Var_C_L, Var_n_C = Var_n_L
  , Var_L_n = Var_L_L, Var_I_n = Var_I_L, Var_N_n = Var_N_L, Var_X_n = Var_X_L, Var_F_n = Var_F_L, Var_f_n = Var_f_L, Var_C_n = Var_C_L, Var_n_n = Var_n_L
  )

dat2 <-
  tibble::tibble(
    Var_d2  = rep("dat2 only", n_obs)
  , Var_L_L = rbinom(n_obs, size = 1, prob = 0.5) |> as.logical()
  , Var_L_I = rbinom(n_obs, size = 1, prob = 0.5) |> as.integer()
  , Var_L_N = rbinom(n_obs, size = 1, prob = 0.5) |> as.numeric()
  , Var_L_X = complex(real = stats::rnorm(n_obs), imaginary = stats::rnorm(n_obs))
  , Var_L_F = sample(letters[ 1: 6], n_obs, replace = TRUE) |> factor(levels = letters[ 1: 6])
  , Var_L_f = sample(letters[21:26], n_obs, replace = TRUE) |> factor(levels = letters[21:26])
  , Var_L_C = sample(letters[11:16], n_obs, replace = TRUE)
  , Var_L_n = rep(NA, n_obs)
  , Var_I_L = Var_L_L, Var_I_I = Var_L_I, Var_I_N = Var_L_N, Var_I_X = Var_L_X, Var_I_F = Var_L_F, Var_I_f = Var_L_f, Var_I_C = Var_L_C, Var_I_n = Var_L_n
  , Var_N_L = Var_L_L, Var_N_I = Var_L_I, Var_N_N = Var_L_N, Var_N_X = Var_L_X, Var_N_F = Var_L_F, Var_N_f = Var_L_f, Var_N_C = Var_L_C, Var_N_n = Var_L_n
  , Var_X_L = Var_L_L, Var_X_I = Var_L_I, Var_X_N = Var_L_N, Var_X_X = Var_L_X, Var_X_F = Var_L_F, Var_X_f = Var_L_f, Var_X_C = Var_L_C, Var_X_n = Var_L_n
  , Var_F_L = Var_L_L, Var_F_I = Var_L_I, Var_F_N = Var_L_N, Var_F_X = Var_L_X, Var_F_F = Var_L_F, Var_F_f = Var_L_f, Var_F_C = Var_L_C, Var_F_n = Var_L_n
  , Var_f_L = Var_L_L, Var_f_I = Var_L_I, Var_f_N = Var_L_N, Var_f_X = Var_L_X, Var_f_F = Var_L_F, Var_f_f = Var_L_f, Var_f_C = Var_L_C, Var_f_n = Var_L_n
  , Var_C_L = Var_L_L, Var_C_I = Var_L_I, Var_C_N = Var_L_N, Var_C_X = Var_L_X, Var_C_F = Var_L_F, Var_C_f = Var_L_f, Var_C_C = Var_L_C, Var_C_n = Var_L_n
  , Var_n_L = Var_L_L, Var_n_I = Var_L_I, Var_n_N = Var_L_N, Var_n_X = Var_L_X, Var_n_F = Var_L_F, Var_n_f = Var_L_f, Var_n_C = Var_L_C, Var_n_n = Var_L_n
  )

# Can't bind by row, variables with same names but different classes
try(dat12 <- dplyr::bind_rows(dat1, dat2))

# align classes
out <- e_data_class_align_between_datasets(dat1, dat2)
out[[ "dat_class_all"     ]] |> print(n=Inf)
out[[ "dat_class_diff_NA" ]] |> print(n=Inf)
out[[ "dat_class_final"   ]] |> print(n=Inf)
out[[ "dat1"              ]] |> str()
out[[ "dat2"              ]] |> str()

# now they can bind by row
dat12 <- dplyr::bind_rows(out$dat1, out$dat2)

}
