% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_complete_multiple_keys_ALL.R
\name{e_data_coalesce_by_column}
\alias{e_data_coalesce_by_column}
\title{Combine rows with different data values}
\usage{
e_data_coalesce_by_column(dat)
}
\arguments{
\item{dat}{data.frame of keys (or other columns) to combine}
}
\value{
coalesced data.frame
}
\description{
For matching \code{group_by()} keys, keeps first non-NA value for each column.
}
\examples{
dat <-
  # A = 1, easy
  # A = 2, easy
  # A = 3, 3rd row has a dup and NA
  # A = 4, 3rd row has conflucing info
  tibble::tribble(
   ~A, ~B, ~C, ~D, ~E
  , 1, NA,  3, NA,  5
  , 1,  2, NA,  2, NA
  , 2, NA, NA,  3, NA
  , 2,  4,  5, NA,  4
  , 3, NA, NA,  3, NA
  , 3,  4,  5, NA,  4
  , 3,  4, NA, NA,  4
  , 4, NA, NA,  3, NA
  , 4,  4,  5, NA,  4
  , 4, 99, 99, 99, NA
  )

dat |>
  dplyr::group_by(A) |>
  dplyr::summarise_all( e_data_coalesce_by_column )
}
