% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_complete_multiple_keys_ALL.R
\name{e_data_replace_keys_less_with_most_complete}
\alias{e_data_replace_keys_less_with_most_complete}
\title{Match and replace less complete rows with most complete row}
\usage{
e_data_replace_keys_less_with_most_complete(
  dat_data,
  dat_most_complete,
  col_keys = c("a", "b", "c")
)
}
\arguments{
\item{dat_data}{data.frame of data for which to update the keys}

\item{dat_most_complete}{data.frame with most complete keys}

\item{col_keys}{names of columns to be used as keys}
}
\value{
data.frame updated with data keys
}
\description{
Match and replace less complete rows with most complete row
}
