% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_data_var_labels_apply_from_table.R
\name{e_data_var_labels_apply_from_table}
\alias{e_data_var_labels_apply_from_table}
\title{Apply labels to a dataset from a table of Var and Label}
\usage{
e_data_var_labels_apply_from_table(dat, var_labels)
}
\arguments{
\item{dat}{a data.frame}

\item{var_labels}{a data.frame with columns \code{Var} and \code{Label}, where \code{Var} are column names that may appear in \code{dat}}
}
\value{
dat the same data.frame with attribute "label" added
}
\description{
Apply labels to a dataset from a table of Var and Label
}
\examples{
dat <- tibble::tibble(a = 1:10, b = 11:20, c = 21:30)
var_labels <-
  tibble::tribble(
    ~Var, ~Label
  , "a" , "Var a Label"
  , "b" , "Var b Label"
  , "c" , "Var c Label"
  )
dat |> str()
dat |> e_data_var_labels_apply_from_table(var_labels) |> str()
}
