% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_data_var_labels_clear.R
\name{e_data_var_labels_clear}
\alias{e_data_var_labels_clear}
\title{Clear labels from variables in data frames}
\usage{
e_data_var_labels_clear(dat)
}
\arguments{
\item{dat}{a data.frame or list of data.frames with attributes "label" to remove}
}
\value{
dat the same object with attribute "label" removed (set to \code{NULL})
}
\description{
Clear labels from variables in data frames
}
\examples{
# a single data.frame
dat <- dat_mtcars_e
dat \%>\% str()
dat \%>\% e_data_var_labels_clear() \%>\% str()
# a list of data.frames
dat <- list(dat_mtcars_e, dat_mtcars_e)
dat \%>\% str()
dat \%>\% e_data_var_labels_clear() \%>\% str()
}
