% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_data_var_labels_table_from_data.R
\name{e_data_var_labels_table_from_data}
\alias{e_data_var_labels_table_from_data}
\title{Create table of labels from a labelled dataset, columns Var and Label}
\usage{
e_data_var_labels_table_from_data(dat)
}
\arguments{
\item{dat}{a data.frame with attribute "label" for columns}
}
\value{
var_labels a data.frame with columns \code{Var} and \code{Label}, where \code{Var} are column names that may appear in \code{dat}
}
\description{
Create table of labels from a labelled dataset, columns Var and Label
}
\examples{
# remove two variable labels
labelled::var_label(dat_mtcars_e[,"vs"]) <- NULL
labelled::var_label(dat_mtcars_e[,"am"]) <- NULL
# get table of labels
var_labels <- dat_mtcars_e |> e_data_var_labels_table_from_data()
var_labels |> print()
}
