% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_dist_test.R
\name{e_dist_test}
\alias{e_dist_test}
\title{Distribution tests}
\usage{
e_dist_test(x = NULL, dist = c("normal")[1])
}
\arguments{
\item{x}{single list of numeric values}

\item{dist}{distribution}
}
\value{
out      list of tests and results
}
\description{
Distribution tests
}
\examples{
rnorm(30) |> e_dist_test()
rexp(30) |> e_dist_test()

}
