% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_distr_test.R
\name{e_distr_test}
\alias{e_distr_test}
\title{Distribution tests}
\usage{
e_distr_test(x = NULL, distr = c("normal")[1])
}
\arguments{
\item{x}{single list of numeric values}

\item{distr}{distribution}
}
\value{
out      list of tests and results
}
\description{
Distribution tests
}
\examples{
rnorm(30) |> e_distr_test()
rexp(30) |> e_distr_test()

}
