% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_SPECIAL_functions_ALL.R
\name{e_duplicated_all}
\alias{e_duplicated_all}
\title{All Determine Duplicate Elements}
\usage{
e_duplicated_all(dat)
}
\arguments{
\item{dat}{list or data.frame}
}
\value{
logical for all duplicated values
}
\description{
\code{e_duplicated_all()} determines which elements of a vector or data frame are duplicates, and returns a logical vector indicating which elements (rows) are duplicates.
}
\examples{
duplicated      (c(1, 2, 2, 3, 4, 5, 2))
e_duplicated_all(c(1, 2, 2, 3, 4, 5, 2))

dat <- data.frame(a = c(1,2,1,2,1,2), b = c(3,3,4,4,3,4))
dat
dat[e_duplicated_all(dat), ]
}
