% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_expand_grid_df.R
\name{e_expand_grid_df}
\alias{e_expand_grid_df}
\title{Create a Data Frame from All Combinations of Data Frames}
\usage{
e_expand_grid_df(...)
}
\arguments{
\item{...}{a list of data.frames}
}
\value{
A data frame containing one block of rows for each combination of
the supplied data frames. The last data frames vary fastest. The columns retain
their original names except for identical names in different data frames, in
which case the suffix ".x", ".y", etc., are added in the returned data frame but may not be unique.
}
\description{
\code{expand.grid} for data.frames.  Create a data frame from all combinations of the supplied data frames.
}
\examples{
df1 <- data.frame(A = 1:3, B = 11:13)
df2 <- data.frame(C = 51:52, D = c("Y", "N"))
df3 <- data.frame(E = c("+", "-"))
e_expand_grid_df(df1, df2, df3)
}
