% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_file_first_last.R
\name{e_file_first_last}
\alias{e_file_first_last}
\title{List of filenames as the first or last n by filename or file time matching a regular expression within a path}
\usage{
e_file_first_last(
  name = NULL,
  path = ".",
  sw_first_last = c("first", "last")[2],
  n_file = 1,
  sw_sort = c("filename", "mtime", "ctime", "atime")[1],
  sw_rev_order = c(TRUE, FALSE)[1],
  ...
)
}
\arguments{
\item{name}{filename regular expression for which to search}

\item{path}{path}

\item{sw_first_last}{first (oldest) or last (newest) set of files}

\item{n_file}{number of filenames to return}

\item{sw_sort}{sort by filename, or by time from \code{file.inf}, file modification, last status change, or last access time}

\item{sw_rev_order}{default sort starting with first (oldest) or reverse to start with last (newest)?}

\item{...}{additional arguments passed to \code{list.files} in addition to \code{path} and \code{pattern}}
}
\value{
fn_out         filename list with path
}
\description{
List of filenames as the first or last n by filename or file time matching a regular expression within a path
}
\examples{
\dontrun{

# last filename by name in getwd()
e_file_first_last()
# newest 3 files in getwd()
e_file_first_last(n_file = 3, sw_sort = "mtime")

}
}
