% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_file_newest_oldest.R
\name{e_file_newest_oldest}
\alias{e_file_newest_oldest}
\title{List of filenames as the first or last n by filename or file time matching a regular expression within a path}
\usage{
e_file_newest_oldest(
  name = NULL,
  path = ".",
  sw_new_old = c("new", "old")[1],
  n_file = 1,
  sw_sort = c("filename", "mtime", "ctime", "atime")[1],
  sw_oldest_first = c(TRUE, FALSE)[1],
  ...
)
}
\arguments{
\item{name}{filename regular expression to search for}

\item{path}{path}

\item{sw_new_old}{newest or oldest files}

\item{n_file}{number of filenames to return}

\item{sw_sort}{sorty by filename, or time to use from \code{file.inf}, file modification, last status change, or last access time.}

\item{sw_oldest_first}{sort oldest first or reverse for newest first?}

\item{...}{additional arguments passed to list.files in addition to \code{path} and \code{pattern}}
}
\value{
fn_out         filename list with path
}
\description{
List of filenames as the first or last n by filename or file time matching a regular expression within a path
}
\examples{
\dontrun{

# last filename by name in getwd()
e_file_newest_oldest()
# newest 3 files in getwd()
e_file_newest_oldest(n_file = 3, sw_sort = "mtime")

}
}
