% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_glm_power.R
\name{e_glm_power}
\alias{e_glm_power}
\title{Logistic Regression power analysis (determine sample size)}
\usage{
e_glm_power(
  p1 = 0.5,
  p2 = 0.5,
  r2 = 0,
  sw_alpha = 0.05,
  sw_power = 0.8,
  sw_print_details = c(TRUE, FALSE)[1],
  sw_glm_type = c("logistic")[1]
)
}
\arguments{
\item{p1}{is the probability that the outcome equals 1 when the predictor x is at its mean.}

\item{p2}{is the probability that the outcome equals 1 when the predictor x is one standard deviation above its mean. Both \code{p1} and \code{p2} must be strictly greater than 0 and strictly less than 1. The best way to obtain these numbers is through a pilot study that closely mimics your research design or from previous literature.}

\item{r2}{is the squared multiple correlation between the predictor variable x and all other variables in the model z1-zk (default = 0).  Use r2=0 for a one-predictor model with only x.}

\item{sw_alpha}{is the significance level for the test. It is the probability of rejecting the null hypothesis when the null hypothesis is true (default = 0.05).}

\item{sw_power}{is the probability of detecting a "true" effect when it exists, defined as 1 minus the probability of incorrectly accepting the null hypothesis (default = 0.80).}

\item{sw_print_details}{T/F, prints an explanation (default = TRUE)}

\item{sw_glm_type}{glm model type, currently only \code{"logistic"}.}
}
\value{
out      a list containing the inputs and the result \code{n}, the sample size required for the logit model
}
\description{
This function computes the sample size required for a logistic regression
with a single continuous covariate of interest x
and one or more optional control variables  z1-zk, Pr(y=1 | x, z1,...zk),
}
\details{
It is based on Phil Ender's powerlog.ado (version 1.0) for Stata,
which is itself based on POWERLOG.SAS (version 1.1) for SAS by Michael Friendly.
\url{https://www.datavis.ca/books/vcd/powerlog.html}
}
\examples{

out_glm_power <-
  e_glm_power(
    p1                = 0.2
  , p2                = 0.5
  , r2                = 0
  , sw_alpha          = 0.05
  , sw_power          = 0.8
  , sw_print_details  = c(TRUE, FALSE)[1]
  , sw_glm_type       = c("logistic")[1]
  )
out_glm_power |> print()

}
