% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_functions_ALL.R
\name{e_health_calc_a1c_categories}
\alias{e_health_calc_a1c_categories}
\title{A1c concentration to CDC classification category labels}
\usage{
e_health_calc_a1c_categories(a1c, sw_fine_cat = c(FALSE, TRUE)[1])
}
\arguments{
\item{a1c}{a1c concentration values}

\item{sw_fine_cat}{T/F Finer classification for A1c category labels. False = 3 categories with cut points: 5.7 and 6.4. True = 5 categories with cut points: 5.7, 6.4, 9, and 12.}
}
\value{
a1c_cat      categorical levels of A1c
}
\description{
https://www.cdc.gov/diabetes/managing/managing-blood-sugar/a1c.html
}
\examples{
0:16 |> e_calc_health_a1c_categories()
0:16 |> e_calc_health_a1c_categories(sw_fine_cat = c(FALSE, TRUE)[2])
}
