% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_functions_ALL.R
\name{e_health_calc_bmi}
\alias{e_health_calc_bmi}
\title{Calculate BMI}
\usage{
e_health_calc_bmi(weight, height, system = c("Metric", "English")[1])
}
\arguments{
\item{weight}{list of weights (kg or lb)}

\item{height}{list of heights (cm or in)}

\item{system}{"Metric" (kg, cm) or "English" (lb, in)}
}
\value{
bmi     value
}
\description{
Calculate BMI
}
\examples{
e_health_calc_bmi(
  weight = 160
, height = 71
, system = c("Metric", "English")[2]
)
}
