% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_functions_ALL.R
\name{e_health_calc_bmi_categories}
\alias{e_health_calc_bmi_categories}
\title{CDC classification for BMI category labels}
\usage{
e_health_calc_bmi_categories(bmi)
}
\arguments{
\item{bmi}{numeric BMI}
}
\value{
bmi_cat, categorical levels of BMI
}
\description{
https://www.cdc.gov/healthyweight/assessing/bmi/adult_bmi/index.html
}
\examples{
e_health_calc_bmi(
  weight = 160
, height = 71
, system = c("Metric", "English")[2]
) |>
e_health_bmi_categories()
}
