% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_functions_ALL.R
\name{e_health_calc_cahps_cat_to_val}
\alias{e_health_calc_cahps_cat_to_val}
\title{CAHPS categorical to numeric response}
\usage{
e_health_calc_cahps_cat_to_val(
  cahps_cat,
  cahps_scale = c("neg1pos1", "other")[1],
  cahps_direction = c("low_neg", "low_pos")[1]
)
}
\arguments{
\item{cahps_cat}{categorical variable from CAHPS}

\item{cahps_scale}{scale type.  "neg1pos1" goes from minimum (negative) and maximum (positive) values for the scale}

\item{cahps_direction}{is low value category associated with negative or positive values}
}
\value{
cahps_val, numeric value of CAHPS
}
\description{
CAHPS categorical to numeric response
}
\examples{
e_health_calc_cahps_cat_to_val(
  cahps_cat       = c(1,2,1,2) |> factor(levels = c(1, 2), labels = c("No", "Yes"))
, cahps_scale     = c("neg1pos1", "other")[1]
, cahps_direction = c("low_neg", "low_pos")[1]
)
}
