% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_functions_ALL.R
\name{e_health_calc_npsy_standardize_scores}
\alias{e_health_calc_npsy_standardize_scores}
\title{Neuropsychological (npsy) scores standardization}
\usage{
e_health_calc_npsy_standardize_scores(
  scores = NULL,
  gender = NULL,
  age = NULL,
  education = NULL,
  sw_scale_name = "MOCATOTS",
  sw_only_print_variable_list = c(FALSE, TRUE)[1]
)
}
\arguments{
\item{scores}{a list of scores to standardize}

\item{gender}{a list of gender in form "M" for male and "F" for female}

\item{age}{a list of ages in years}

\item{education}{a list of education in years, e.g., 12 indicates completed high school}

\item{sw_scale_name}{a text string naming the scale to standarize, case insensitive}

\item{sw_only_print_variable_list}{T/F to print the available \code{sw_scale_name} options}
}
\value{
scores_standardized          a list of standardized scores
}
\description{
Neuropsychological (npsy) scores standardization
}
\examples{
e_health_calc_npsy_standardize_scores(
    sw_only_print_variable_list = c(FALSE, TRUE)[2]
  )

e_health_calc_npsy_standardize_scores(
    scores        = rep(rep(seq(5, 50, by = 5), 6), 2)
  , gender        = rep(rep(c(rep("M", 10), rep("F", 10)), 3), 2)
  , age           = rep(c(rep(30, 20), rep(50, 20), rep(70, 20)), 2)
  , education     = c(rep(10, 60), rep(16, 60))
  , sw_scale_name = "MOCATOTS"
  , sw_only_print_variable_list = c(FALSE, TRUE)[1]
  )

# Below values should result in these values:
val_true <- c(1.35, -0.38, -7.48, 1.08, 0.49, -0.37, -0.58, -2.56, -0.6, -2.86, -1.41)
val_stand <-
  e_health_calc_npsy_standardize_scores(
    scores        = c(29, 26, 6, 29, 28, 25, 25, 19, 23, 18, 22)
  , gender        = c("F", "F", "F", "F", "F", "F", "F", "F", "F", "M", "F")
  , age           = c(72, 55, 77, 72, 70, 64, 79, 67, 87, 83, 65)
  , education     = c(14, 14, 12, 16, 17, 13, 18, 12, 14, 16, 12)
  , sw_scale_name = "MOCATOTS"
  , sw_only_print_variable_list = c(FALSE, TRUE)[1]
  ) |>
  round(2)
val_stand - val_true
}
