% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_functions_ALL.R
\name{e_health_calc_phq2_categories}
\alias{e_health_calc_phq2_categories}
\title{PHQ-2 Scores and Proposed Treatment Actions, Table 4 from Kroenke K, Spitzer RL, Psychiatric Annals 2002;32:509-521}
\usage{
e_health_calc_phq2_categories(phq2_total_score)
}
\arguments{
\item{phq2_total_score}{total score}
}
\value{
phq2_cat, categorical levels of PHQ-2
}
\description{
PHQ-2 Scores and Proposed Treatment Actions, Table 4 from Kroenke K, Spitzer RL, Psychiatric Annals 2002;32:509-521
}
\examples{
0:6 |> e_health_calc_phq2_categories()
}
