% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_health_functions_ALL.R
\name{e_health_calc_poverty_ratio_FederalPovertyLevel}
\alias{e_health_calc_poverty_ratio_FederalPovertyLevel}
\title{Federal Poverty Level ratio}
\usage{
e_health_calc_poverty_ratio_FederalPovertyLevel(
  income_annual = 0,
  n_persons = 1,
  poverty_year = 2016
)
}
\arguments{
\item{income_annual}{annual income in dollars}

\item{n_persons}{number of persons in household}

\item{poverty_year}{year for comparison (currently only for 2016)}
}
\value{
poverty_ratio, the ratio of a person's income to the Federal Poverty Level based on the number of people in the household
}
\description{
Federal Poverty Level ratio
}
\examples{
e_health_calc_poverty_ratio_FederalPovertyLevel(
    income_annual = seq(0, 60000, by = 10000)
  , n_persons     = 1:7
  , poverty_year  = 2016
)
}
