% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_is_error.R
\name{e_is_error}
\alias{e_is_error}
\title{check a function and if there's an error return an indication that we can ifelse() to either run for real or skip}
\usage{
e_is_error(f)
}
\arguments{
\item{f}{}
}
\value{

}
\description{
check a function and if there's an error return an indication that we can ifelse() to either run for real or skip
}
\examples{
sw_try_ok <-
  !e_is_error(
    try(
      # try what you want, check for error
      log(NULL)
    )
  )
if(sw_try_ok) {
  # no error, do what you intended to do
} else {
  # error, do something else
}
}
