% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_kable_latex.R
\name{e_kable_latex}
\alias{e_kable_latex}
\title{Title}
\usage{
e_kable_latex(dat, sw_scale = FALSE, sw_kable_format = c("html", "latex")[1])
}
\arguments{
\item{sw_kable_format}{}
}
\value{

}
\description{
Title
}
