% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_lm_bestsubset.R
\name{e_lm_bestsubset}
\alias{e_lm_bestsubset}
\title{Best subset selection returns results sorted by BIC}
\usage{
e_lm_bestsubset(form, dat, nbest = 5)
}
\arguments{
\item{form}{lm formula for full model}

\item{dat}{data to use}

\item{nbest}{number of models to return for each model size, default is 5}
}
\value{
tibble (data.frame) of selection results
}
\description{
Best subset selection returns results sorted by BIC
}
\examples{
\dontrun{
tab_best <-
  e_lm_bestsubset(
    form = formula(mpg ~ cyl + disp + hp + gear)
  , dat  = datasets::mtcars
  )

  op <- options(); # saving old options
  options(width=100) # setting command window output text width wider
tab_best \%>\% print(n = Inf, width = Inf)
  options(op); # reset (all) initial options
}
}
