% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_lm_diag_plots.R
\name{e_lm_diag_plots}
\alias{e_lm_diag_plots}
\title{Plotting residual diagnostics for an lm() object.}
\usage{
e_lm_diag_plots(
  fit = NULL,
  rc_mfrow = c(1, 3),
  which_plot = c(4, 6, 1),
  n_outliers = 3,
  sw_qqplot = TRUE,
  sw_boxcox = TRUE,
  sw_constant_var = TRUE,
  sw_collinearity = TRUE,
  sw_order_of_data = TRUE,
  sw_addedvar = TRUE,
  sw_plot_set = c(NA, "simple", "simpleAV", "all")[1]
)
}
\arguments{
\item{fit}{linear model object returned by lm()}

\item{rc_mfrow}{number of rows and columns for the graphic plot, default is c(1, 3); use "NA" for a single plot with 3 columns}

\item{which_plot}{default plot numbers for lm()}

\item{n_outliers}{number to identify in plots from lm() and qqPlot()}

\item{sw_qqplot}{T/F for whether to show the QQ-plot}

\item{sw_boxcox}{T/F for whether to show Box-Cox transformation}

\item{sw_constant_var}{T/F for whether to assess constant variance}

\item{sw_collinearity}{T/F for whether to assess multicollinearity between predictor variables}

\item{sw_order_of_data}{T/F for whether to show residuals by order of data}

\item{sw_addedvar}{T/F for whether to show added-variables plot}

\item{sw_plot_set}{NULL to accept other plot options, or "simple" to exclude boxcox, constant var, collinearity, order of data, and added-variable plots. "simpleAV" to add back in the added-variable plots.  "all" includes all possible plots in this function.}
}
\value{
NULL, invisibly
}
\description{
Plotting residual diagnostics for an lm() object.
}
\examples{
fit <- lm(mpg ~ cyl + disp + hp + gear, data = datasets::mtcars)
e_lm_diag_plots(fit)
mod <- formula(mpg ~ cyl + disp + hp + gear)
fit <- lm(mod, data = datasets::mtcars)
e_lm_diag_plots(fit)
}
