% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_lm_model_criteria.R
\name{e_lm_model_criteria}
\alias{e_lm_model_criteria}
\title{Return table of model selection criteria}
\usage{
e_lm_model_criteria(lm_fit = NULL, dat_fit = NULL, model_id = NULL)
}
\arguments{
\item{lm_fit}{fitted model object}

\item{dat_fit}{data used for model fit}

\item{model_id}{label for each model that is fit, helps to match up with original model}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a tibble of model selection criteria
}\if{html}{\out{</div>}}
}
\description{
Calculates a large set of model selection criteria and returns them in a one-row tibble.

AIC and BIC criteria from the stats package:
\itemize{
\item \code{aic} Akaike's An Information Criterion
\item \code{bic} Bayesian Information Criterion (AIC with k = log(nobs(object)))
\item \code{caic} Conditional Akaike Information for 'lme4' and 'lme'
\item \code{nobs} number of observations used to fit \code{lm_fit}
}

Several criteria from the lm summary:
\itemize{
\item \code{r2} R-squared statistic
\item \code{r2adj} Adjusted R-squared statistic
\item \code{f_stat} F-statistic compared to grand mean model
\item \code{p} F-statistic numerator degrees-of-freedom (number of parameters in the model)
\item \code{df} F-statistic denominator degrees-of-freedom (number of df estimating the variance)
\item \code{f_stat_pval} F-statistic p-value
\item \code{rse} Residual standard error.
}

Several criteria are defined in \code{?modelr::`model-quality`}:

Three summaries are immediately interpretible on the scale of the response
variable:
\itemize{
\item \code{rmse} is the root-mean-squared-error
\item \code{mae} is the mean absolute error
\item \code{qae} is quantiles of absolute error.
}

Other summaries have varying scales and interpretations:
\itemize{
\item \code{mape} mean absolute percentage error.
\item \code{rsae} is the relative sum of absolute errors.
\item \code{mse} is the mean-squared-error.
\item \code{rsquare} is the variance of the predictions divided by the
variance of the response.
}
}
\examples{
lm_form <- formula(mpg ~ cyl + carb + disp + hp + disp:hp + wt + vs + am + gear)
lm_fit <-
  stats::lm(
    formula = lm_form
  , data    = dat_mtcars_e
  )
lm_crit <-
  e_lm_model_criteria(
    lm_fit   = lm_fit
  , dat_fit  = dat_mtcars_e
  , model_id = 1
  )
lm_crit
}
