% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_log_shift.R
\name{e_log_shift}
\alias{e_log_shift}
\title{Log2 nonnegative numbers after offsetting 0 by minimum non-zero value.}
\usage{
e_log_shift(x, min_add = NULL, base_log = 2)
}
\arguments{
\item{x}{numeric vector}

\item{min_add}{value to add to x before taking log, if not specified and min(x) <= 0, then this is determined automatically}

\item{base_log}{base for log}
}
\value{
numeric vector with attributes
}
\description{
Default \code{min_add} values:
  If specified, use that.
  All positive, add 0.
  Min is 0, add the minimum non-0 value.
  Min is negative and max is positive, add the minimum non-0 value (so minimum is now the previous non-0 value).
  Max is negative, add the minimum value plus 1 so \code{log2(x)=0} is the minimum value.
}
\examples{

e_log_shift(x = c(0, 1, 2, 3), min_add = 1)
e_log_shift(x = c(0, 10, 100, 1000), base = 10)
e_log_shift(x = c(-4, -2, 0, 2, 4))
}
