% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_log_write.R
\name{e_log_write}
\alias{e_log_write}
\title{Initialize and print into a log file and to the console}
\usage{
e_log_write(
  log_text = NULL,
  log_obj = log_obj,
  i_level = 2,
  sw_init = c(FALSE, TRUE)[1],
  out_path = ".",
  file_prefix = "out"
)
}
\arguments{
\item{log_text}{text string to print}

\item{log_obj}{log_obj object that includes filename, console function, and \code{time_start}}

\item{i_level}{Numbers 1 to 5 corresponding to "DEBUG", "INFO", "WARN", "ERROR", "FATAL" from \code{log4r::loglevel()}s}

\item{sw_init}{TRUE to initilize log file and console logging, FALSE to print text}

\item{out_path}{out path for log file}

\item{file_prefix}{filename prefix for log file}
}
\value{
log_obj or NULL
}
\description{
Initialize and print into a log file and to the console
}
\examples{
\dontrun{

# initialize log file
log_obj <-
  e_log_write(
    sw_init     = TRUE
  , out_path    = "."
  , file_prefix = "out"
  )
# write log entry
e_log_write(
    log_text    = "Write this message"
  , log_obj     = log_obj
  , i_level     = 2
  )

}
}
