% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_mnv_norm_qqplot.R
\name{e_mnv_norm_qqplot}
\alias{e_mnv_norm_qqplot}
\title{Graphical Assessment (QQ-plot) for assessing Multivariate Normality}
\usage{
e_mnv_norm_qqplot(x, name = "")
}
\arguments{
\item{x}{data.frame or matrix of numeric columns}

\item{name}{label for title}
}
\value{

}
\description{
Graphical Assessment (QQ-plot) for assessing Multivariate Normality
}
\examples{
e_mnv_norm_qqplot(shells[shells$sex == "F", 2:4], "Female")
e_mnv_norm_qqplot(shells[shells$sex == "M", 2:4], "Male")
}
