% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_model_binary_complete.R
\name{e_model_binary_complete}
\alias{e_model_binary_complete}
\title{All variable subsets}
\usage{
e_model_binary_complete(var_names)
}
\arguments{
\item{var_names}{a character list of variable names}
}
\value{
a data.frame of all binary subsets of variables
}
\description{
Return data.frame of all binary subsets of a list of variable names
}
\examples{

e_model_binary_complete(
   var_names = c("a", "b", "c")
 )

}
