% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_model_selection.R
\name{e_model_selection}
\alias{e_model_selection}
\title{Model selection function for lm and glm, using \code{step} via AIC/BIC or best subset.}
\usage{
e_model_selection(
  form = NULL,
  dat = NULL,
  sw_model = c("lm", "glm")[1],
  sw_sel_type = c("step", "bestsubset")[1],
  sw_step_direction = c("both", "backward", "forward")[1],
  sw_step_k = c("AIC", "BIC", 2)[1],
  sw_glm_scale = c("link", "response")[2],
  sw_plot_missing = c(TRUE, FALSE)[1],
  sw_plot_y_covar = c(TRUE, FALSE)[1],
  sw_plot_x_corr = c(TRUE, FALSE)[1],
  sw_contrasts = c("both", "sel", "none", "skip")[2],
  sw_print_results = c(TRUE, FALSE)[1],
  emmip_rg.limit = 10000,
  ...
)
}
\arguments{
\item{form}{formula for a model}

\item{dat}{data to use}

\item{sw_model}{type of regression model, \code{"lm"} or \code{"glm"}}

\item{sw_sel_type}{type of model selection.  \code{step} starts with specified \code{form} model with upper = two-way effects model and lower = 1 (intercept-only) model}

\item{sw_step_direction}{for \code{step}, the \code{direction} argument from the \code{stats::step()} function}

\item{sw_step_k}{for \code{step} determined by "AIC" (2), "BIC" (log(n)), or specified numerically; the \code{k} argument from the \code{stats::step()} function for penalty df}

\item{sw_glm_scale}{Scale for glm contrasts, \code{"link"} or \code{"response"}, when \code{sw_model} is \code{"glm"}}

\item{sw_plot_missing}{output plots from \code{e_plot_missing}}

\item{sw_plot_y_covar}{output plots from \code{e_plot_lm_y_covar}}

\item{sw_plot_x_corr}{output correlation matrix plot}

\item{sw_contrasts}{which contrasts to plot, \code{"both"} is full and selected model, \code{"sel"} is only selected model, \code{"none"} produces blank plots (but this may still take a while due to large grid with many interactions), and \code{"skip"} doesn't run \code{e_plot_model_contrasts()}.}

\item{sw_print_results}{print results before returning object}

\item{...}{options sent to \code{e_model_all_subsets_formula} for subsets to consider}
}
\value{
out                list object with all results organized by \code{init} and \code{sel} for initial and selected models
}
\description{
(Best subset not yet implemented.)
}
\examples{
\dontrun{
## lm example
dat_sel <-
   erikmisc::dat_mtcars_e
form_model <-
  mpg ~ cyl + disp + hp + wt + vs + am + cyl:vs + disp:hp + hp:vs

out <-
  e_model_selection(
    form = form_model
  , dat  = dat_sel
  )
# out |> print()


## glm example
dat_sel <-
  erikmisc::dat_mtcars_e |>
  dplyr::mutate(
    am_01 =
      dplyr::case_when(
        am == "manual"    ~ 0
      , am == "automatic" ~ 1
      )
  )
labelled::var_label(dat_sel[["am_01"]]) <- labelled::var_label(dat_sel[["am"]])
form_model <-
  cbind(am_01, 1 - am_01) ~ cyl + disp + hp + wt + vs + hp:vs

out <-
  e_model_selection(
    form       = form_model
  , dat        = dat_sel
  , sw_model   = c("lm", "glm")[2]
  , sw_step_k  = c("AIC", "BIC", 2)[2]
  )
# out |> print()
}
}
