% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_HELPERS.R
\name{e_plot_blank}
\alias{e_plot_blank}
\title{Create blank ggplot grob plot used as a place holder}
\usage{
e_plot_blank(text_label = c("", "N/A")[1])
}
\arguments{
\item{text_label}{optional text label, such as "N/A"}
}
\value{
p_blank blank ggplot grob, possibly with a text label
}
\description{
Create blank ggplot grob plot used as a place holder
}
\examples{
p_blank <- e_plot_blank(text_label = c("", "N/A")[2])
}
