% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_bs_one_samp_dist.R
\name{e_plot_bs_one_samp_dist}
\alias{e_plot_bs_one_samp_dist}
\title{Visual comparison of whether Bootstrap sampling distribution is close to Normal}
\usage{
e_plot_bs_one_samp_dist(dat, N = 10000)
}
\arguments{
\item{dat}{a list of values}

\item{N}{number of bootstrap iterations}
}
\value{
\code{invisible(NULL)}
}
\description{
A function to compare the bootstrap sampling distribution with
  a normal distribution with mean and SEM estimated from the data
}
\examples{
e_plot_bs_one_samp_dist(dat = runif(6))
}
