% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_missing.R
\name{e_plot_calc_break_interval}
\alias{e_plot_calc_break_interval}
\title{Interval for plot breaks, determine a "nice" length of seq(by = Interval)}
\usage{
e_plot_calc_break_interval(
  values,
  num_intervals = 10,
  val_leading_digit = c(1, 2, 5, 10)
)
}
\arguments{
\item{values}{numeric values for an axis}

\item{num_intervals}{rough desired number of intervals}

\item{val_leading_digit}{leading digits to consider (in ascending order, end with 10)}
}
\value{
interval for breaks
}
\description{
Interval for plot breaks, determine a "nice" length of seq(by = Interval)
}
\examples{
e_plot_calc_break_interval(1:250)
e_plot_calc_break_interval(1:1000)
e_plot_calc_break_interval(-1:1)
}
