% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_clean_numeric_nonnum_minmax.R
\name{e_plot_clean_numeric_nonnum_minmax}
\alias{e_plot_clean_numeric_nonnum_minmax}
\title{Cleaning numeric values for nonnumeric and large/small values, similarly for dates.}
\usage{
e_plot_clean_numeric_nonnum_minmax(
  dat_this,
  binwidth = 7,
  n_table = 5,
  sw_date = FALSE,
  sw_plot = TRUE,
  lab_title = NULL
)
}
\arguments{
\item{dat_this}{data gathered with selected variables}

\item{binwidth}{for histogram}

\item{n_table}{is the number of table values to print}

\item{sw_date}{is the value a date?}

\item{sw_plot}{TRUE/FALSE switch to plot}

\item{lab_title}{title before output and for plot}
}
\value{
invisible(NULL)
}
\description{
The variables included together are plotted on the same x-axis scale.
}
\examples{
## requires formatting into two columns: "variables" and "values"
dat_this <-
  dat_mtcars_e |>
  dplyr::select(
    model
  , tidyselect::starts_with("hp")
  , where(is.numeric)
  ) |>
  tidyr::pivot_longer(
    cols      = -model
  , names_to  = "variable"
  , values_to = "value"
  ) |>
  dplyr::select(
    -model
  )
dat_this |> e_plot_clean_numeric_nonnum_minmax()
}
