% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_compare_medians.R
\name{e_plot_compare_medians}
\alias{e_plot_compare_medians}
\title{ANOVA-style dot-plot comparing means}
\usage{
e_plot_compare_medians(
  dat_plot,
  var_response = NULL,
  var_groups = NULL,
  label_response = NULL,
  label_groups = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  comparisons = c("all", list(), NULL)[1],
  cm_method = c("wilcox.test", "t.test")[1],
  cm_p.adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none")[1],
  cm_dot = c("mean", "median")[2],
  cm_fun = c("mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median",
    "median_iqr", "median_hilow", "median_q1q3", "median_mad", "median_range")[7],
  cm_error.plot = c("pointrange", "linerange", "crossbar", "errorbar",
    "upper_errorbar", "lower_errorbar", "upper_pointrange", "lower_pointrange",
    "upper_linerange", "lower_linerange")[4],
  sw_add = c("none", "dotplot", "jitter", "boxplot", "point")[2],
  sw_caption_desc = c(TRUE, FALSE)[1],
  symnum.args = list(cutpoints = c(0, 1e-04, 0.001, 0.01, 0.05, 0.1, 1), symbols =
    c("****", "***", "**", "*", "-", "ns"))
)
}
\arguments{
\item{dat_plot}{Data.frame}

\item{var_response}{Variable name of response (y) (as string, "var_name")}

\item{var_groups}{Variable name of groups (x) (as string, "var_name")}

\item{label_response}{Label for y-axis}

\item{label_groups}{Label for x-axis}

\item{plot_title}{Plot title}

\item{plot_subtitle}{Plot subtitle}

\item{comparisons}{NULL=no comparisons, "all"=all pairwise comparisons, or a list of comparisons to make, such as \code{list(c("A", "B"), c("A", "C"))}.}

\item{cm_method}{Compare method: "wilcox.test" for nonparametric, "t.test" for parametric}

\item{cm_p.adjust.method}{Compare method: p-value adjust method "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{cm_dot}{Compare method: red center estimate "mean", "median"}

\item{cm_fun}{Compare method: red error bar range "mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median", "median_iqr", "median_hilow", "median_q1q3", "median_mad", "median_range"}

\item{cm_error.plot}{Compare method: red error bar style "pointrange", "linerange", "crossbar", "errorbar", "upper_errorbar", "lower_errorbar", "upper_pointrange", "lower_pointrange", "upper_linerange", "lower_linerange"}

\item{sw_add}{Add a second plot to the violin plot: "none", "dotplot", "jitter", "boxplot", "point"}

\item{sw_caption_desc}{T/F for caption describing red summary center and spread}

\item{symnum.args}{List of cutpoints and symbols for significance}
}
\value{
p_pub              \code{p_pub$result_compare_means} has pairwise comparison table.  \code{p_pub} is a ggplot plot grob.
}
\description{
Some functions won't work without running \code{library(ggpubr)}.
}
\examples{
# medians
e_plot_compare_medians(
    dat_plot            = dat_mtcars_e
  , var_response        = "mpg"
  , var_groups          = "cyl"
  , label_response      = labelled::var_label(dat_mtcars_e$mpg)
  , label_groups        = labelled::var_label(dat_mtcars_e$cyl)
  , plot_title          = "MPG by cylinders"
  , plot_subtitle       = "Comparison of medians"
  , comparisons         = "all"
  , cm_method           = "wilcox.test"
  , cm_p.adjust.method  = "holm"
  , cm_dot              = "median"
  , cm_fun              = "median_iqr"
  , cm_error.plot       = "errorbar"
  , sw_add              = "dotplot"
  , sw_caption_desc     = TRUE
  )

# means
e_plot_compare_medians(
    dat_plot            = dat_mtcars_e
  , var_response        = "mpg"
  , var_groups          = "cyl"
  , label_response      = labelled::var_label(dat_mtcars_e$mpg)
  , label_groups        = labelled::var_label(dat_mtcars_e$cyl)
  , plot_title          = "MPG by cylinders"
  , plot_subtitle       = "Comparison of meeans"
  , comparisons         = "all"
  , cm_method           = "t.test"
  , cm_p.adjust.method  = "holm"
  , cm_dot              = "mean"
  , cm_fun              = "mean_ci"
  , cm_error.plot       = "errorbar"
  , sw_add              = "boxplot"
  , sw_caption_desc     = FALSE
  )

}
