% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_corr_ellipse.R
\name{e_plot_corr_ellipse}
\alias{e_plot_corr_ellipse}
\title{Correlation plot with ellipses}
\usage{
e_plot_corr_ellipse(
  corr,
  outline = FALSE,
  col = "grey",
  upper.panel = c("ellipse", "number", "none")[1],
  lower.panel = c("ellipse", "number", "none")[2],
  diag = c("none", "ellipse", "number")[3],
  digits = 2,
  bty = "n",
  axes = FALSE,
  xlab = "",
  ylab = "",
  asp = 1,
  cex.lab = par("cex.lab"),
  cex = 0.75 * par("cex"),
  mar = c(0, 0, 1, 0),
  ...
)
}
\arguments{
\item{corr}{correlation matrix}

\item{outline}{option}

\item{col}{option}

\item{upper.panel}{option}

\item{lower.panel}{option}

\item{diag}{option}

\item{digits}{option}

\item{bty}{option}

\item{axes}{option}

\item{xlab}{option}

\item{ylab}{option}

\item{asp}{option}

\item{cex.lab}{option}

\item{cex}{option}

\item{mar}{option}

\item{...}{option}
}
\value{
\code{invisible(NULL)}
}
\description{
Correlation plot with ellipses
}
\examples{
datasets::mtcars |> cor() |> e_plot_corr_ellipse()

\dontrun{
## Color version
# use mtcars since all numeric
tab_cor <-
  datasets::mtcars |>
  as.matrix() |>
  Hmisc::rcorr(
    type = c("pearson", "spearman")[1]
  )
tab_cor$r

# red neg, white 0, blue pos
color_scale <-
  c(
    rgb(241,  54,  23, maxColorValue = 255)
  , "white"
  , rgb(  0,  61, 104, maxColorValue = 255)
  )
# color ramp in "Lab" space
color_ramp <- colorRampPalette(color_scale, space = "Lab")
# Create scale with 200 points
colors_plot <- color_ramp(100)

# move correlation range from [-1, 1] to [0, 100] for colors
tab_cor$r |>
e_plot_corr_ellipse(
  col  = colors_plot[round(((tab_cor$r + 1)/2) * 100)]
, main = "Colored ellipses"
, upper.panel = c("ellipse", "number", "none")[1]
, lower.panel = c("ellipse", "number", "none")[2]
, diag = c("none", "ellipse", "number")[3]
, digits = 3
)
}
}
