% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_corr_matrix.R
\name{e_plot_corr_matrix}
\alias{e_plot_corr_matrix}
\title{Correlation matrix plot with ellipses}
\usage{
e_plot_corr_matrix(
  dat = NULL,
  sw_type = c("pearson", "spearman")[1],
  sw_conf_level = 0.95,
  sw_plot_type = c("full", "mixed")[1],
  sw_cluster_rect = NULL,
  sw_hclust_method = c("ward", "ward.D", "ward.D2", "single", "complete", "average",
    "mcquitty", "median", "centroid")[3]
)
}
\arguments{
\item{dat}{data.frame or tibble to plot a correlation matrix of just the numeric variables.}

\item{sw_type}{Pearson or Spearman correlation}

\item{sw_conf_level}{Confidence level for significance, typically 0.95.}

\item{sw_plot_type}{A full symmetric plot or a mixed plot with different visualizations on upper and lower triangle.}

\item{sw_cluster_rect}{Number of cluster rectangles, available for \code{sw_plot_type = "full"}}

\item{sw_hclust_method}{Clustering method for sorting variables and grouping with rectangles.}
}
\value{
out        a list, including tables of correlations and p-values from \code{Hmisc::rcorr}, CIs for r from \code{corrplot::cor.mtest}, and a plot from \code{corrplot::corrplot} as a ggplot grob
}
\description{
See \code{corrplot::corrplot()} for creating custom versions of this
}
\examples{
erikmisc::dat_mtcars_e |> e_plot_corr_matrix()

out <- erikmisc::dat_mtcars_e |> e_plot_corr_matrix(sw_cluster_rect = 3)
out$plot |> print()

out <- erikmisc::dat_mtcars_e |> e_plot_corr_matrix(sw_plot_type = "mixed")
out$plot |> print()
}
