% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_function_limit_range.R
\name{e_plot_function_limit_range}
\alias{e_plot_function_limit_range}
\title{Limit the range of values from a distribution function}
\usage{
e_plot_function_limit_range(fun = dnorm, min = -Inf, max = +Inf, ...)
}
\arguments{
\item{fun}{function, such as distribution function like \code{dnorm}}

\item{min}{min value to include in range}

\item{max}{max value to include in range}

\item{...}{arguments passed to fun}
}
\value{
filtered range of the function
}
\description{
Limit the range of values from a distribution function
}
\examples{

f_dnorm_limit <- e_plot_function_limit_range(dnorm, 0, 2)
f_dnorm_limit(-2:4)
p <- ggplot(data.frame(x = c(-3, 3)), aes(x = x))
p <- p + stat_function(fun = dnorm)
p <- p + stat_function(fun = e_plot_function_limit_range(dnorm, 0, 2)
                     , geom = "area", fill = "blue", alpha = 0.2)
}
