% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_grid_corner_label_ALL.R
\name{e_plot_grid_corner_label}
\alias{e_plot_grid_corner_label}
\title{For corner labels, remove the labs(tag=) argument and use this}
\usage{
e_plot_grid_corner_label(
  p,
  label_corner = "A",
  font_color = "black",
  font_size = 18,
  font_face = "bold"
)
}
\arguments{
\item{p}{ggplot grob}

\item{label_corner}{label for top-left corner of plot}

\item{font_color}{color}

\item{font_size}{point size}

\item{font_face}{font face}
}
\value{
grob with top-left corner label
}
\description{
See \code{?e_plot_grid_corner_label_list} to use with a list.
}
