% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_longitudinal.R
\name{e_plot_longitudinal}
\alias{e_plot_longitudinal}
\title{Longitudinal interaction plots}
\usage{
e_plot_longitudinal(
  dat_plot = dat,
  var_x_time = "x",
  var_y_resp = "y",
  var_ID = "ID",
  var_group = NULL,
  label_x_time = NULL,
  label_y_resp = NULL,
  label_group = NULL,
  x_scale_breaks = NULL,
  y_scale_breaks = NULL,
  sw_group_reverse = FALSE,
  hist_binwidth = NULL,
  hist_align = c("center", "boundary")[1],
  line_type_grand = c("none", "solid", "dashed", "dotted", "dotdash", "longdash",
    "twodash")[6],
  line_type_group = c("none", "solid", "dashed", "dotted", "dotdash", "longdash",
    "twodash")[3]
)
}
\arguments{
\item{line_type_group}{}
}
\value{

}
\description{
Longitudinal interaction plots
}
\examples{
data("sleepstudy")

f_plot_longitudinal(
    dat_plot         = sleepstudy
  , var_x_time       = "Days"
  , var_y_resp       = "Reaction"
  , var_ID           = "Subject"
  , var_group        = NULL
  , label_x_time     = "Number of days of sleep deprivation"
  , label_y_resp     = "Average reaction time (ms)"
  #, label_group      = "Program"
  , x_scale_breaks   = seq(0, 9, by = 2)
  , y_scale_breaks   = seq(0, 1000, by = 100)
  #, sw_group_reverse = TRUE
  , hist_binwidth    = 25
  , hist_align       = c("center", "boundary")[2]
  , line_type_grand  = c("none", "solid", "dashed", "dotted", "dotdash", "longdash", "twodash")[6]
  , line_type_group  = c("none", "solid", "dashed", "dotted", "dotdash", "longdash", "twodash")[3]
  )
}
