% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_missing.R
\name{e_plot_missing}
\alias{e_plot_missing}
\title{Plots missing data in a data.frame, possibly grouped by one variable and sorted by a second.}
\usage{
e_plot_missing(dat, var_group = NULL, sw_group_sort = FALSE, var2_sort = NULL)
}
\arguments{
\item{dat}{data.frame or tibble}

\item{var_group}{variable name to group by (colors data)}

\item{sw_group_sort}{TRUE/FALSE to sort by grouped variable}

\item{var2_sort}{second variable name to sort by if data is grouped}
}
\value{
ggplot grob plot object
}
\description{
Plots missing data in a data.frame, possibly grouped by one variable and sorted by a second.
}
\examples{
dat = datasets::mtcars
prop_missing = 0.10
n_missing = sample.int(n = prod(dim(dat)), size = round( prop_missing * prod(dim(dat))))
ind_missing = expand.grid(1:dim(dat)[1], 1:dim(dat)[2])[n_missing, ]
for (i_row in seq_along(n_missing)) {
  dat[ind_missing[i_row,1], ind_missing[i_row,2] ] <- NA
}

e_plot_missing(
    dat            = dat
  , var_group      = "cyl"
  , sw_group_sort  = TRUE
  , var2_sort      = "disp"
  )
}
