% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_mnv_norm_qqplot.R
\name{e_plot_mnv_norm_qqplot}
\alias{e_plot_mnv_norm_qqplot}
\title{Graphical Assessment (QQ-plot) for assessing Multivariate Normality}
\usage{
e_plot_mnv_norm_qqplot(x, name = "")
}
\arguments{
\item{x}{data.frame or matrix of numeric columns}

\item{name}{label for title}
}
\value{
\code{invisible(NULL)}
}
\description{
Graphical Assessment (QQ-plot) for assessing Multivariate Normality
}
\examples{
datasets::mtcars \%>\%
  dplyr::filter(cyl == 4) \%>\%
  dplyr::select(mpg, hp, wt) \%>\%
  e_plot_mnv_norm_qqplot(name = "cyl = 4")
datasets::mtcars \%>\%
  dplyr::filter(cyl == 6) \%>\%
  dplyr::select(mpg, hp, wt) \%>\%
  e_plot_mnv_norm_qqplot(name = "cyl = 6")
datasets::mtcars \%>\%
  dplyr::filter(cyl == 8) \%>\%
  dplyr::select(mpg, hp, wt) \%>\%
  e_plot_mnv_norm_qqplot(name = "cyl = 8")
}
