% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics.R
\name{e_plot_model_diagnostics}
\alias{e_plot_model_diagnostics}
\title{Plot and report model diagnostics for a lm or glm object.}
\usage{
e_plot_model_diagnostics(fit = NULL, dat = NULL, ...)
}
\arguments{
\item{fit}{model object}

\item{dat}{data used for model fit
\subsection{@param sw_plot_set        NULL to accept other plot options, or "simple" to exclude boxcox, constant var, collinearity, order of data, and added-variable plots. "simpleAV" to add back in the added-variable plots.  "all" includes all possible plots in this function.}{
}

\subsection{@param rp_type            option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_bins            option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_smoother        option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_qqline          option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_qqbands         option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_scale           option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_theme           option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_axis.text.size  option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_title.text.size option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_title.opt       option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param rp_alpha           option passed to \code{ggResidpanel::resid_panel}}{
}

\subsection{@param ...                options passed to other function.}{
}

\subsection{@param rc_mfrow           number of rows and columns for the graphic plot, default is c(1, 3); use "NA" for a single plot with 3 columns}{
}

\subsection{@param which_plot         default plot numbers for lm()}{
}

\subsection{@param n_outliers         number to identify in plots from lm() and qqPlot()}{
}

\subsection{@param sw_qqplot          T/F for whether to show the QQ-plot}{
}

\subsection{@param sw_boxcox          T/F for whether to show Box-Cox transformation}{
}

\subsection{@param sw_constant_var    T/F for whether to assess constant variance}{
}

\subsection{@param sw_collinearity    T/F for whether to assess multicollinearity between predictor variables}{
}

\subsection{@param sw_order_of_data   T/F for whether to show residuals by order of data}{
}

\subsection{@param sw_addedvar        T/F for whether to show added-variables plot}{
}}
}
\value{
out_diagn        list of tables and plots
}
\description{
Many are created from package \code{ggResidpanel}, working with models of
types \code{lm}, \code{glm}, \code{lme}, \code{lmer}, \code{glmer},
and \code{lmerTest}.
}
\details{
The following options can be chosen for the \code{sw_plot_set} argument.
\itemize{
\item "all": This creates a panel of all plot types included in the package
that are available for the model type input into \code{resid_panel}. (See note
below.)
\item "default": This creates a panel with a residual plot, a normal quantile plot
of the residuals, an index plot of the residuals, and a histogram of the residuals.
\item "R": This creates a panel with a residual plot, a normal
quantile plot of the residuals, a location-scale plot, and a residuals versus leverage
plot. This was modeled after the plots shown in R if the \code{plot()} base function
is applied to an \code{lm} model. This option can only be used with an \code{lm} or
\code{glm} model.
\item "SAS": This creates a panel with a residual plot, a normal quantile plot of
the residuals, a histogram of the residuals, and a boxplot of the residuals.
This was modeled after the residualpanel option in proc mixed from SAS version 9.4.
\item A vector of individual plots can also be specified.
For example, one can specify \code{plots = c("boxplot", "hist")} or
\code{plots = "qq"}. The individual plot options are as follows.
\itemize{
\item \code{"boxplot"}: A boxplot of residuals
\item \code{"cookd"}: A plot of Cook's D values versus observation numbers
\item \code{"hist"}: A histogram of residuals
\item \code{"index"}: A plot of residuals versus observation numbers
\item \code{"ls"}: A location scale plot of the residuals
\item \code{"qq"}: A normal quantile plot of residuals
\item \code{"lev"}: A plot of standardized residuals versus leverage values
\item \code{"resid"}: A plot of residuals versus predicted values
\item \code{"yvp":}: A plot of observed response values versus predicted values
} }
}
\examples{
## lm example
dat <-
   erikmisc::dat_mtcars_e
form_model_lm <-
  mpg ~ cyl + disp + hp + wt + vs + am + disp:hp + hp:vs
fit_lm <- lm(formula = form_model_lm, data = dat)
out_diagn <- e_plot_model_diagnostics(fit = fit_lm, dat = dat)

## glm example
dat <-
  erikmisc::dat_mtcars_e |>
  dplyr::mutate(
    am_01 =
      dplyr::case_when(
        am == "manual"    ~ 0
      , am == "automatic" ~ 1
      )
  )
labelled::var_label(dat[["am_01"]]) <- labelled::var_label(dat[["am"]])
form_model_glm <-
  cbind(am_01, 1 - am_01) ~ cyl + disp + hp + wt + vs + hp:vs
fit_glm <- glm(formula = form_model_glm, data = dat, family = binomial(link = "logit"))
out_diagn <- e_plot_model_diagnostics(fit = fit_glm, dat = dat)

}
