% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_AllTestTables_combine}
\alias{e_plot_model_diagnostics_AllTestTables_combine}
\title{Model diagnostics, Combine all test tables}
\usage{
e_plot_model_diagnostics_AllTestTables_combine(
  out_diagn = out_diagn,
  sw_interp = c(TRUE, FALSE)[2]
)
}
\arguments{
\item{out_diagn}{final object from \cdoe{e_plot_model_diagnostics()}}

\item{sw_interp}{T/F to provide interpretation guidance in the plot caption or with a table.}
}
\value{
out      table of all diagnostic numeric tests
}
\description{
Model diagnostics, Combine all test tables
}
