% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__avPlots}
\alias{e_plot_model_diagnostics_car__avPlots}
\title{Model diagnostics, Added-variable plots, car::avPlots}
\usage{
e_plot_model_diagnostics_car__avPlots(
  fit = NULL,
  sw_avplot_main_only = c(TRUE, FALSE)[1]
)
}
\arguments{
\item{fit}{fit object}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, Added-variable plots, car::avPlots
}
