% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__boxCox}
\alias{e_plot_model_diagnostics_car__boxCox}
\title{Model diagnostics, car::boxCox and car::symbox}
\usage{
e_plot_model_diagnostics_car__boxCox(fit = NULL)
}
\arguments{
\item{fit}{fit object}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, car::boxCox and car::symbox
}
