% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__ceresPlots}
\alias{e_plot_model_diagnostics_car__ceresPlots}
\title{Model diagnostics, Ceres (Generalized Partial Residual) Plots, car::ceresPlots}
\usage{
e_plot_model_diagnostics_car__ceresPlots(fit = NULL)
}
\arguments{
\item{fit}{fit object}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, Ceres (Generalized Partial Residual) Plots, car::ceresPlots
}
