% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__infIndexPlot}
\alias{e_plot_model_diagnostics_car__infIndexPlot}
\title{Model diagnostics, Influence index plot, car::infIndexPlot}
\usage{
e_plot_model_diagnostics_car__infIndexPlot(fit = NULL)
}
\arguments{
\item{fit}{fit object}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, Influence index plot, car::infIndexPlot
}
