% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__inverseResponsePlot}
\alias{e_plot_model_diagnostics_car__inverseResponsePlot}
\title{Model diagnostics, car::inverseResponsePlot}
\usage{
e_plot_model_diagnostics_car__inverseResponsePlot(fit = NULL)
}
\arguments{
\item{fit}{fit object}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, car::inverseResponsePlot
}
