% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__residualPlots}
\alias{e_plot_model_diagnostics_car__residualPlots}
\title{Model diagnostics, resid vs x vars, car::residualPlots}
\usage{
e_plot_model_diagnostics_car__residualPlots(
  fit = NULL,
  resid_type = c("working", "response", "deviance", "pearson", "partial", "rstudent",
    "rstandard")[6]
)
}
\arguments{
\item{fit}{fit object}

\item{resid_type}{Specifies the type of residual to be plotted.  Any of \code{c("working", "response", "deviance", "pearson", "partial", "rstudent", "rstandard")} may be specified.  The default \code{resid_type = "pearson"} is usually appropriate, since it is equal to ordinary residuals observed minus fit with ols, and correctly weighted residuals with wls or for a glm.  The last two options use the \code{\link{rstudent}} and \code{\link{rstandard}} functions and use studentized or standardized residuals.}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, resid vs x vars, car::residualPlots
}
