% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__spreadLevelPlot}
\alias{e_plot_model_diagnostics_car__spreadLevelPlot}
\title{Model diagnostics, car::spreadLevelPlot and car::ncvTest}
\usage{
e_plot_model_diagnostics_car__spreadLevelPlot(fit = NULL)
}
\arguments{
\item{fit}{fit object}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, car::spreadLevelPlot and car::ncvTest
}
