% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_CooksD_Leverage_Resid}
\alias{e_plot_model_diagnostics_CooksD_Leverage_Resid}
\title{Model diagnostics, Cooks vs Leverage vs Residuals}
\usage{
e_plot_model_diagnostics_CooksD_Leverage_Resid(
  fit = NULL,
  fit_resid = NULL,
  fit_cooksD = NULL,
  fit_leverage = NULL,
  sw_interp = c(TRUE, FALSE)[2]
)
}
\arguments{
\item{fit}{fit object}

\item{fit_resid}{list of residuals from \code{e_model_calc_resid()}}

\item{fit_cooksD}{Cook's Distance values (from \code{stats::cooks.distance})}

\item{fit_leverage}{leverage values (from \code{stats::hatvalues})}

\item{sw_interp}{T/F to provide interpretation guidance in the plot caption or with a table.}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, Cooks vs Leverage vs Residuals
}
