% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_DHARMa_Resid}
\alias{e_plot_model_diagnostics_DHARMa_Resid}
\title{Model diagnostics, DHARMa: residual diagnostics for hierarchical (multi-level/mixed) regression models}
\usage{
e_plot_model_diagnostics_DHARMa_Resid(fit = NULL, dat = NULL)
}
\arguments{
\item{fit}{fit object}

\item{dat}{dataset data.frame or tibble}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, DHARMa: residual diagnostics for hierarchical (multi-level/mixed) regression models
}
