% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_qqplotr}
\alias{e_plot_model_diagnostics_qqplotr}
\title{Model diagnostics, qqplot, qqplotr:: stat_* functions}
\usage{
e_plot_model_diagnostics_qqplotr(
  fit_resid = fit_resid,
  fit = NULL,
  band_conf = 0.95,
  sw_interp = c(TRUE, FALSE)[2]
)
}
\arguments{
\item{fit_resid}{list of residuals from \code{e_model_calc_resid()}}

\item{fit}{fit object}

\item{band_conf}{confidence level of confidence band}

\item{sw_interp}{T/F to provide interpretation guidance in the plot caption or with a table.}
}
\value{
out      list including ggplot grobs, one normal, one detrended
}
\description{
Model diagnostics, qqplot, qqplotr:: stat_* functions
}
