% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_Resid_histogram}
\alias{e_plot_model_diagnostics_Resid_histogram}
\title{Model diagnostics, Cooks vs Leverage vs Residuals}
\usage{
e_plot_model_diagnostics_Resid_histogram(
  fit_resid = NULL,
  sw_interp = c(TRUE, FALSE)[2]
)
}
\arguments{
\item{fit_resid}{list of residuals from \code{e_model_calc_resid()}}

\item{sw_interp}{T/F to provide interpretation guidance in the plot caption or with a table.}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, Cooks vs Leverage vs Residuals
}
