% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_projection.R
\name{e_plot_projection}
\alias{e_plot_projection}
\title{Plot projection from high dimensions to 2D}
\usage{
e_plot_projection(
  dat_plot = NULL,
  var_group = NULL,
  var_color = NULL,
  var_shape = NULL,
  var_facet = NULL,
  text_title = NULL,
  sw_print = FALSE,
  sw_projection = c("umap", "tsne")[1],
  n_obs_sample = NULL
)
}
\arguments{
\item{dat_plot}{data to plot}

\item{var_group}{variable name to exclude from projection, but may be used for plotting}

\item{var_color}{variable color}

\item{var_shape}{variable shape}

\item{var_facet}{variable to by which to facet wrap}

\item{text_title}{text for title of plot to override default}

\item{sw_print}{T/F whether to print plot}

\item{sw_projection}{Projection type: "tsne" (others to come)}

\item{n_obs_sample}{number of observations to sample from data since projection calculation can be very expensive for large samples (only for tsne)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a ggplot object
}\if{html}{\out{</div>}}
}
\description{
UMAP is very fast and seems to separate better than slower t-SNE.
}
\examples{
e_plot_projection(
    dat_plot                  = dat_mtcars_e |> dplyr::select(mpg, cyl, disp, hp, vs)
  , var_group                 = NULL
  , var_color                 = NULL
  , var_shape                 = NULL
  , var_facet                 = NULL
  , text_title                = NULL
  , sw_print                  = FALSE
  , sw_projection             = c("umap", "tsne")[1]
  , n_obs_sample              = NULL
  )

e_plot_projection(
    dat_plot                  = dat_mtcars_e |> dplyr::select(mpg, cyl, disp, hp, vs)
  , var_group                 = "cyl"
  , var_color                 = "cyl"
  , var_shape                 = "cyl"
  , var_facet                 = "vs"
  , text_title                = NULL
  , sw_print                  = FALSE
  , sw_projection             = c("umap", "tsne")[1]
  , n_obs_sample              = NULL
  )

}
