% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_rf_vimp.R
\name{e_plot_rf_vimp}
\alias{e_plot_rf_vimp}
\title{A function to plot the "all" column from a Random Forests Variable Importance Plot (VIMP).}
\usage{
e_plot_rf_vimp(v_imp, targets = NULL)
}
\arguments{
\item{v_imp}{\code{$importance} table from \code{randomForestSRC::rfsrc()}}

\item{targets}{a list of column names of \code{v_imp}, such as \code{c("all", "Yes")}.  If \code{NULL}, does "all" and every target.}
}
\value{
p
}
\description{
A function to plot the "all" column from a Random Forests Variable Importance Plot (VIMP).
}
\examples{
# o_class <- randomForestSRC::rfsrc(...)
# e_plot_rf_vimp(o_class$importance)
}
