% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_roc_reorder_hierarchy.R
\name{e_plot_roc_reorder_hierarchy}
\alias{e_plot_roc_reorder_hierarchy}
\title{A helper function for a list of ROC curves to reorder ROC objects by type (rather than target)}
\usage{
e_plot_roc_reorder_hierarchy(list_roc = NULL)
}
\arguments{
\item{list_roc}{list of ROC curves from \code{e_plot_roc} ordered by target}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       list of ROC curves reorder ROC objects by type (rather than target)
}\if{html}{\out{</div>}}
}
\description{
A helper function for a list of ROC curves to reorder ROC objects by type (rather than target)
}
\examples{
# See e_plot_roc()

}
