% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_scatterplot.R
\name{e_plot_scatterplot}
\alias{e_plot_scatterplot}
\title{Plot scatterplot of two numeric variables, by up to one color factor variables and two facets factor variables}
\usage{
e_plot_scatterplot(
  dat_plot = NULL,
  var_num = NULL,
  var_color = NULL,
  var_facet = NULL,
  sw_print = FALSE
)
}
\arguments{
\item{dat_plot}{data to plot}

\item{var_num}{two numeric variable names to plot}

\item{var_color}{factor or numeric color variable}

\item{var_facet}{factor varibles (1 or 2) to facet by, (row facets, then column facets),}

\item{sw_print}{T/F whether to print table and display plot}
}
\value{
\preformatted{     a list including a table and ggplot
}
}
\description{
Plot scatterplot of two numeric variables, by up to one color factor variables and two facets factor variables
}
\examples{
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_num   = c("mpg", "disp")
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_num   = c("mpg", "disp")
  , var_color = c("cyl")
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_num   = c("mpg", "disp")
  , var_color = c("wt")          # numeric color
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_num   = c("mpg", "disp")
  , var_facet = c("vs", "am")
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_num   = c("mpg", "disp")
  , var_color = c("cyl")
  , var_facet = c("vs", "am")
  )
}
