% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_table_prop.R
\name{e_plot_table_prop}
\alias{e_plot_table_prop}
\title{Calculate proportions and plot factor variable by up to two other factor variables}
\usage{
e_plot_table_prop(dat_plot = NULL, var_names = NULL, sw_print = FALSE)
}
\arguments{
\item{dat_plot}{data to plot}

\item{var_names}{proportions are calculated for first variable, by up to two other variables (row facets, then column facets), \code{c("y", "a", "b")}}

\item{sw_print}{T/F whether to print table and display plot}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a list including a table and ggplot object
}\if{html}{\out{</div>}}
}
\description{
Calculate proportions and plot factor variable by up to two other factor variables
}
\examples{
e_plot_table_prop(
    dat_plot  = dat_mtcars_e
  , var_names = c("vs")
  )
e_plot_table_prop(
    dat_plot  = dat_mtcars_e
  , var_names = c("vs", "am")
  )
e_plot_table_prop(
    dat_plot  = dat_mtcars_e
  , var_names = c("vs", "am", "gear")
  )
}
