% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_print_list_tree_hierarchy.R
\name{e_print_list_tree_hierarchy}
\alias{e_print_list_tree_hierarchy}
\title{Print list tree hierarchy}
\usage{
e_print_list_tree_hierarchy(x = NULL)
}
\arguments{
\item{x}{list of lists}
}
\value{
list_tree Tree as a data.frame
}
\description{
runs \code{data.tree::FromListSimple()}
}
\examples{
x <- list()
x$aa <- list()
x$aa <- list(x = list(xx = 1), y = list(yy = 2), z = list(zz = 3))
x$bb <- list()
x$bb <- list(xx = 4, yy = 5, zz = 6)
e_print_list_tree_hierarchy(x)
}
