% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_pval_empirical.R
\name{e_pval_empirical}
\alias{e_pval_empirical}
\title{Calculate empirical tail probabilities (p-values), typically of bootstrap resamples}
\usage{
e_pval_empirical(
  x = NULL,
  obs = 0,
  prefix_pval = "pval_",
  suffix_tail = c("lower", "upper", "twoside", "min_uplo")
)
}
\arguments{
\item{x}{vector of values}

\item{obs}{observed value to calculate tail probabilities}

\item{prefix_pval}{prefix for p-value named list}

\item{suffix_tail}{suffixes for p-value named list}
}
\value{
out        a named vector of 4 (pval_lower, pval_upper, pval_twoside, pval_min_uplo)
}
\description{
Calculate empirical tail probabilities (p-values), typically of bootstrap resamples
}
\examples{
out = e_pval_empirical(x = rnorm(1000), obs = 2, prefix_pval = "this_pval_")
out |> print()
}
