% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_read_data_files.R
\name{e_read_data_files}
\alias{e_read_data_files}
\title{Read a list of csv, xls, or xlsx data into a list of tibbles}
\usage{
e_read_data_files(
  read_fn_path = ".",
  read_fn_names = NULL,
  sw_dat_add_col_path_fn = c(TRUE, FALSE)[1],
  sw_dat_print_fn_read = c(TRUE, FALSE)[1]
)
}
\arguments{
\item{read_fn_path}{Directory path}

\item{read_fn_names}{Character list of file names}

\item{sw_dat_add_col_path_fn}{T/F for data, add two columns specifying the directory (\code{DIR__}) and filename (\code{FILE__})}

\item{sw_dat_print_fn_read}{T/F print file names and dimensions as the files are read}
}
\value{
dat_sheet             A list of tibbles
}
\description{
Called from \code{e_read_data_subdir_into_lists()}.
}
\examples{
\dontrun{
# See e_read_data_subdir_into_lists()
}
}
