% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_read_data_subdir_into_lists.R
\name{e_read_data_subdir_into_lists}
\alias{e_read_data_subdir_into_lists}
\title{Read data from subdirectories into a structured list}
\usage{
e_read_data_subdir_into_lists(
  fn_path = ".",
  fn_detect = c("csv$", "xls$", "xlsx$"),
  sw_fn_or_dat = c("fn", "dat")[1],
  sw_exclude_empty_dir = c(TRUE, FALSE)[1],
  sw_dat_add_col_path_fn = c(TRUE, FALSE)[1],
  sw_dat_print_fn_read = c(TRUE, FALSE)[2]
)
}
\arguments{
\item{fn_path}{Starting directory path}

\item{fn_detect}{NULL for all.  File specification, used by \code{stringr::str_detect()}, usually specifying file extensions.}

\item{sw_fn_or_dat}{Return filenames or tibbles (data.frames)}

\item{sw_exclude_empty_dir}{T/F exclude empty directories}

\item{sw_dat_add_col_path_fn}{T/F for data, add two columns specifying the directory (\code{DIR__}) and filename (\code{FILE__})}

\item{sw_dat_print_fn_read}{T/F print file names and dimensions as the files are read}
}
\value{
fn_names              Either a structured list of filenames or of tibbles
}
\description{
Recurse through subdirectories returning either filenames or tibbles (data.frames) by reading csv, xls, or xlsx with \code{e_read_data_files()}.
}
\examples{
\dontrun{
# # all file names
# e_read_data_subdir_into_lists(
#     fn_path   = "./data-raw/dat_subdir"
#   , fn_detect = NULL
#   , sw_fn_or_dat  = c("fn", "dat")[1]
#   , sw_exclude_empty_dir = c(TRUE, FALSE)[1]
#   )
# # selected file names
# e_read_data_subdir_into_lists(
#     fn_path   = "./data-raw/dat_subdir"
#   , fn_detect = c("csv$", "xls$", "xlsx$")
#   , sw_fn_or_dat  = c("fn", "dat")[1]
#   , sw_exclude_empty_dir = c(TRUE, FALSE)[1]
#   )
# # selected data
# e_read_data_subdir_into_lists(
#     fn_path                 = "./data-raw/dat_subdir"
#   , fn_detect               = c("csv$", "xls$", "xlsx$")
#   , sw_fn_or_dat            = c("fn", "dat")[2]
#   , sw_exclude_empty_dir    = c(TRUE, FALSE)[1]
#   , sw_dat_add_col_path_fn  = c(TRUE, FALSE)[1]
#   , sw_dat_print_fn_read    = c(TRUE, FALSE)[1]
#   )
}
}
