% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_REDCap_read_format_data.R
\name{e_REDCap_read_format_data}
\alias{e_REDCap_read_format_data}
\title{REDCap data, reading and formatting}
\usage{
e_REDCap_read_format_data(
  fn_path = getwd(),
  fn_root = NULL,
  fn_suffix_data = "_DATA_",
  fn_suffix_R = "_R_",
  fn_suffix_out = "",
  fn_ext_data = "*.csv",
  fn_ext_R = "*.r",
  sw_return_dat = TRUE,
  sw_save_RData = TRUE
)
}
\arguments{
\item{fn_path}{is the location of the data, assumed to be the current path.}

\item{fn_root}{the root name(s) of the file(s), typically the project name(s) (the part before _DATA_*.*)}

\item{fn_suffix_data}{typically "_DATA_" to indicate this is the data}

\item{fn_suffix_R}{typically "_R_" to indicate this is the R formatting file}

\item{fn_suffix_out}{an added name to the outputted saved and formatted data files, maybe you want to add a date, a version number, or a another label}

\item{fn_ext_data}{"*.csv" to specify the data filename extension, only csv is supported ("csv" is differnt from "CSV")}

\item{fn_ext_R}{"*.r" to specify the R filename extension, ("r" is differnt from "R")}

\item{sw_return_dat}{TRUE/FALSE to return the data object from the function}

\item{sw_save_RData}{TRUE/FALSE to save an ".RData" file}
}
\value{

}
\description{
REDCap data, reading and formatting
}
\examples{
read_format_data(fn_path = NULL) # for Erik to update package data
read_format_data()            # for others to format data in current path
}
