% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_REDCap_replace_var_label.R
\name{e_REDCap_replace_var_label}
\alias{e_REDCap_replace_var_label}
\title{REDCap variable labeling, replace Hmisc \code{label()} with labelled \code{var_label()} in the R formatting file}
\usage{
e_REDCap_replace_var_label(dat_to_var_label = NULL)
}
\arguments{
\item{dat_to_var_label}{is the R file that needs \code{label()} replaced with \code{var_label()}}
}
\value{
file with \code{var_label()} replaced
}
\description{
REDCap variable labeling, replace Hmisc \code{label()} with labelled \code{var_label()} in the R formatting file
}
