% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_rfsrc_classification.R
\name{e_rfsrc_classification}
\alias{e_rfsrc_classification}
\title{Random Forests classification workflow using \code{randomForestSRC}}
\usage{
e_rfsrc_classification(
  dat_rf_class = NULL,
  rf_y_var = NULL,
  rf_x_var = NULL,
  sw_rfsrc_ntree = 2000,
  sw_alpha = 0.05
)
}
\arguments{
\item{dat_rf_class}{data.frame with data}

\item{rf_y_var}{y factor variable, if NULL then assumed to be the first column of \code{dat_rf_class}}

\item{rf_x_var}{x variables, if NULL then assumed to be all except the first column of \code{dat_rf_class}}

\item{sw_rfsrc_ntree}{\code{ntree} argument for \code{randomForestSRC::rfsrc()}}

\item{sw_alpha}{\code{alpha} argument for \code{randomForestSRC::plot.subsample()} and \code{randomForestSRC::extract.bootsample()}}
}
\value{
list with many RF objects, summaries, and plots
}
\description{
Random Forests classification workflow using \code{randomForestSRC}
}
\examples{
dat_rf_class <-
  erikmisc::dat_mtcars_e |>
  dplyr::select(
    -model
  ) |>
  dplyr::select(
    cyl
  , tidyselect::everything()
  )

out_e_rf <-
  e_rfsrc_classification(
    dat_rf_class    = dat_rf_class
  , rf_y_var        = NULL
  , rf_x_var        = NULL
  , sw_rfsrc_ntree  = 200
  , sw_alpha        = 0.05
  )

}
