% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_rmd_name_chunks.R
\name{e_rmd_name_chunks}
\alias{e_rmd_name_chunks}
\title{Read an Rmd rmarkdown file and apply enumerated code chunks}
\usage{
e_rmd_name_chunks(fn_in = NULL, fn_out = NULL, prefix_chunk_name = "chunk-")
}
\arguments{
\item{fn_in}{Rmd filename in.}

\item{fn_out}{Rmd filename out.  If NULL, then this is assigned \code{fn_in}.}

\item{prefix_chunk_name}{a string before each chunk number}
}
\value{
NULL, invisibly
}
\description{
Read an Rmd rmarkdown file and apply enumerated code chunks
}
\examples{
\dontrun{
e_rmd_name_chunks(fn_in = "test_in.Rmd", fn_out = "test_out.Rmd", prefix_chunk_name = "chunk-")
}
}
