% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_roc_gg.R
\name{e_roc_gg}
\alias{e_roc_gg}
\title{A function to calculate ROC curve, calculate area under the curve (AUC), calculate the optimal threshold, plot the curve, and classify the peaks.}
\usage{
e_roc_gg(
  dat,
  var_class,
  val_feat,
  sw_plot = TRUE,
  method_name = "",
  sw_thresh_criteria = c("min_dist", "spec1_sensmax", "sens1_specmax")[1]
)
}
\arguments{
\item{dat}{dataset name}

\item{var_class}{true classification}

\item{val_feat}{XXX DON'T REMEMBER}

\item{sw_plot}{TRUE/FALSE switch to create plot or not}

\item{method_name}{PREVIOUSLY USED FOR TITLE}

\item{sw_thresh_criteria}{(min distand, max spec, or max sens)}
}
\value{
out is a list including the optimal threshold, val_feat (XXX), good_class (XXX), plot_roc is a plot of the roc curve using ggplot, and roc_dat includes the curve, best value of the curve, and cm.feat (XXX?)
}
\description{
A function to calculate ROC curve, calculate area under the curve (AUC), calculate the optimal threshold, plot the curve, and classify the peaks.
}
