% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_scale_align.R
\name{e_scale_align}
\alias{e_scale_align}
\title{Align the distribution of one empirical distribution to an empirical reference distribution}
\usage{
e_scale_align(
  var_to_scale = NULL,
  var_ref = NULL,
  scale_method = c("quantile", "zscore")[1],
  quantiles_to_match = c(0.1, 0.9),
  sw_plot = c(TRUE, FALSE)[2]
)
}
\arguments{
\item{var_to_scale}{Variable to rescale}

\item{var_ref}{Reference variable distribution}

\item{scale_method}{Method of rescaling.  \code{"quantile"} with \code{c(0.1, 0.9)} is remarkably good.}

\item{quantiles_to_match}{For \code{"quantile"}, which quantiles to match.}

\item{sw_plot}{T/F to plot Reference, Original, and Scaled variables.}
}
\value{
val_scaled        Scaled vector (NAs have been removed)
}
\description{
Align the distribution of one empirical distribution to an empirical reference distribution
}
\examples{
set.seed(76543)
e_scale_align(
  var_to_scale        = rgamma( 40, shape = 3, rate = 2) + 4
, var_ref             = rgamma(100, shape = 3, rate = 1)
, scale_method        = c("quantile", "zscore")[1]
, quantiles_to_match  = c(0.1, 0.9)
, sw_plot             = c(TRUE, FALSE)[1]
)
set.seed(76543)
e_scale_align(
  var_to_scale        = rgamma( 40, shape = 3, rate = 2) + 4
, var_ref             = rgamma(100, shape = 3, rate = 1)
, scale_method        = c("quantile", "zscore")[2]
, sw_plot             = c(TRUE, FALSE)[1]
)
}
