% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_source_all_R_in_folder.R
\name{e_source_all_R_in_folder}
\alias{e_source_all_R_in_folder}
\title{Source all *.R files in a folder}
\usage{
e_source_all_R_in_folder(fn_R_location = "./R", fn_list_name = "*.R")
}
\arguments{
\item{fn_R_location}{directory (folder) location}

\item{fn_list_name}{file name pattern}
}
\value{
NULL, invisibly
}
\description{
Source all *.R files in a folder
}
