% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_split_list_columns_into_indicator_columns.R
\name{e_split_list_columns_into_indicator_columns}
\alias{e_split_list_columns_into_indicator_columns}
\title{Split a (set of) item-listing columns into indicator columns}
\usage{
e_split_list_columns_into_indicator_columns(
  dat_this,
  var_names_items = NULL,
  item_delimiters = ",.;/|",
  code_other_below_freq = 5,
  label_other = "other",
  indicator_col_prefix = "item_",
  sw_replace_GT1_with_1 = FALSE,
  sw_print_unique = TRUE
)
}
\arguments{
\item{dat_this}{entire data.frame or tibble, will return with additional indicator columns}

\item{var_names_items}{column names with lists of items}

\item{item_delimiters}{delimiter(s) that separate items within a single column}

\item{code_other_below_freq}{replace item name with \code{label_other} if total frequency for an item is less than this value}

\item{label_other}{label for the "other" category}

\item{indicator_col_prefix}{prefix for the new indicator columns}

\item{sw_replace_GT1_with_1}{T/F, to replace "greater than 1" counts with an indicator of 1 (to interpret as "at least 1")}

\item{sw_print_unique}{T/F, print list of items before and after replacing with "other"}
}
\value{
dat_this with additional indicator columns
}
\description{
Commonly used for comorbidities or prescription lists within a single (or multiple) column(s).
Takes a column where items are separated by puncutation (,.;/|) and creates separate columns with indicators.
Can treat counts of items >1 as 1 to simplify summary tables (for example, multiple items coded as "other").
}
\examples{
dat_ex <-
  dplyr::tibble(
    col1 =
      c(
        NA, "", "a", "A, B  ,C", "b", "D. c ;    d"
      , "x  ;Y", "ab/0|J;1;1", 2, "other", 1
      , "a a a", "1,a a a", "2,a a a"
      )
  , col2 = LETTERS[1:length(col1)]
  ) \%>\%
  dplyr::mutate(
    ID = 1:dplyr::n()
  ) \%>\%
  dplyr::select(
    ID
  , everything()
  )
dat_ex \%>\% print(n = Inf)

dat_ex_out <-
  e_split_list_columns_into_indicator_columns(
    dat_this              = dat_ex
  , var_names_items       = c("col1", "col2")
  , item_delimiters       = ",.;/|"
  , code_other_below_freq = 2
  , label_other           = "other"
  , indicator_col_prefix  = "item_"
  , sw_replace_GT1_with_1 = FALSE
  , sw_print_unique       = TRUE
  )
dat_ex_out \%>\% print(n = Inf)
}
