% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_substr_last_n.R
\name{e_substr_last_n}
\alias{e_substr_last_n}
\title{Substring last n characters}
\usage{
e_substr_last_n(string = NULL, last_n = 1)
}
\arguments{
\item{string}{list of strings}

\item{last_n}{number of characters from the end to keep}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   same list of strings but only up to last n characters from each
}\if{html}{\out{</div>}}
}
\description{
Uses \code{stringr::str_sub(string, start = -n)}.
}
\examples{

c("abcdefghij", "This has a space", "abc", NA) |>
 e_substr_last_n(last_n = 6)
}
