% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_tab_sum.R
\name{e_tab_sum}
\alias{e_tab_sum}
\title{e_tab_sum create a frequency/proportion table by a set of variables}
\usage{
e_tab_sum(dat, var_names, sw_sort_prop = TRUE)
}
\arguments{
\item{dat}{data.frame or tibble}

\item{var_names}{list of variable names}

\item{sw_sort_prop}{TRUE/FALSE to sort the last variable descending by proportion within the other variables}
}
\value{
tab_summary a summary table
}
\description{
e_tab_sum create a frequency/proportion table by a set of variables
}
\examples{
e_tab_sum(
    dat           = ggplot2::mpg
  , var_names     = c("drv", "cyl", "class")
  , sw_sort_prop  = TRUE
  ) \%>\%
  print(n = Inf)
}
