% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_table1_summaries.R
\name{e_table1_summaries}
\alias{e_table1_summaries}
\title{Table 1 demographics summaries}
\usage{
e_table1_summaries(
  dat_tab1 = dat_sub,
  list_var_col_table = NULL,
  list_var_col_names = NULL,
  list_var_row_table = NULL,
  list_var_row_names = NULL,
  fn_root = "tab1_vars",
  label_width = 40,
  sw_verbose = FALSE,
  moonbook_digits = c(factor = 1, numeric = 3),
  moonbook_method = c(1, 2, 3)[2]
)
}
\arguments{
\item{dat_tab1}{Dataset for table.}

\item{list_var_col_table}{Column variable name.}

\item{list_var_col_names}{Column variable label.}

\item{list_var_row_table}{Variable names to loop over.}

\item{list_var_row_names}{Variable labels to label table.}

\item{fn_root}{Filename for output table.}

\item{label_width}{Width to wrap column and row variable labels, does not affect category labels.}

\item{sw_verbose}{Print status to console during execution?}

\item{moonbook_digits}{An integer indicating the number of decimal places (round) or significant digits to be used. Default value is 3. (\code{?moonBook::mytable_sub})}

\item{moonbook_method}{An integer indicating methods for continuous variables. Possible values in methods are: 1 = forces analysis as normal-distributed; 2 = forces analysis as continuous non-normal; 3 = performs a Shapiro-Wilk test to decide between normal or non-normal. (\code{?moonBook::mytable_sub})}
}
\value{

}
\description{
Moonbook significant digit version
function to run each variable individually, determining significant digits
then join them into a single file at the end
}
\details{
Uses \code{moonBook::mytable}; default is CatMethod = 2: Numeric summaries are Median and IQR with kruskal.test, Categorical summaries are proportions with chisq.test with continuity correction.
}
