% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_table_aggregate.R
\name{e_table_aggregate}
\alias{e_table_aggregate}
\title{Aggregate a set of variables by another set of variables with a specific function}
\usage{
e_table_aggregate(
  dat,
  var_names_by,
  var_names_aggregate,
  func_aggregate = list(sum),
  .groups = c("drop", "keep")[1]
)
}
\arguments{
\item{dat}{data.frame or tibble}

\item{var_names_by}{list of variable names to group by}

\item{var_names_aggregate}{list of variable names to aggregate}

\item{func_aggregate}{a function (e.g., \code{sum}) to aggregate all variables by, or a list of functions to aggregate each variable by}

\item{.groups}{drops or keeps groups}
}
\value{
aggregated data
}
\description{
Aggregate a set of variables by another set of variables with a specific function
}
\examples{
e_table_aggregate(
    dat                 = dat_mtcars_e
  , var_names_by        = c("cyl")
  , var_names_aggregate = c("disp", "hp", "drat")
  , func_aggregate      = mean
  )
e_table_aggregate(
    dat                 = dat_mtcars_e
  , var_names_by        = c("cyl", "am")
  , var_names_aggregate = c("disp", "hp", "drat", "wt", "qsec")
  , func_aggregate      = list(sum, sum, mean, mean, sum)
  )

# assign variables and functions in a list
list_aggregate_var_func <-
  list(
    `disp` = sum
  , `hp`   = sum
  , `drat` = mean
  , `wt`   = mean
  , `qsec` = sum
  )
e_table_aggregate(
    dat                 = dat_mtcars_e
  , var_names_by        = c("cyl", "am")
  , var_names_aggregate = list_aggregate_var_func |> names()
  , func_aggregate      = list_aggregate_var_func
  )
}
