% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_table_print.R
\name{e_table_print}
\alias{e_table_print}
\title{Print a data.frame table for latex and scale to fit page}
\usage{
e_table_print(
  dat,
  sw_scale = FALSE,
  sw_kable_format = c("simple", "kbl", "html", "latex", "doc")[2],
  sw_latex_options = c("basic", "striped", "hold_position", "HOLD_position",
    "scale_down", "repeat_header")[c(2, 3)],
  ...
)
}
\arguments{
\item{dat}{data.frame or tibble to print}

\item{sw_scale}{if "latex" T/F to scale to fit page, if "html" then a point size to pass to \code{kableExtra::kable_styling} argument \code{font_size}}

\item{sw_kable_format}{"simple", "html", "latex", or "doc"}

\item{sw_latex_options}{passed to \code{kableExtra::kable_styling(latex_options = sw_latex_options)}, updated if \code{sw_scale=TRUE} to include \code{"scale_down"}}

\item{...}{other arguments passed to \code{knitr::kable()}}
}
\value{
\code{invisible(NULL)}
}
\description{
If using in Rmd file and a latex documnet, then requires two header includes:
}
\details{
\preformatted{
EXAMPLE OF .Rmd YAML header for pdf via LaTeX

---
title: "Title"
date: "April 10, 2022"
output:
  pdf_document:
    df_print: kable
    latex_engine: lualatex
    toc: true
    number_sections: true
    toc_depth: 2
    keep_tex: true
header-includes:
- \usepackage{booktabs}
- \usepackage{colortbl}
always_allow_html: yes
fig_caption: TRUE
---
}
}
\examples{
datasets::mtcars \%>\% head() \%>\% e_table_print()
\dontrun{
# html rescaled size
datasets::mtcars \%>\%
  head() \%>\%
  e_table_print(sw_scale = 6)
# outputs into LaTeX document, scaling is automatic to fit page width
# works best before a \clearpage
# use chunk option: results = 'asis', see other options at https://yihui.org/knitr/options/
datasets::mtcars \%>\%
  head() \%>\%
  e_table_print(sw_scale = TRUE, sw_kable_format = "latex")
}
}
