% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_table_sum_freq_prop.R
\name{e_table_sum_freq_prop}
\alias{e_table_sum_freq_prop}
\title{Create a frequency/proportion summary table by a set of variables}
\usage{
e_table_sum_freq_prop(dat, var_names, sw_sort_prop = TRUE)
}
\arguments{
\item{dat}{data.frame or tibble}

\item{var_names}{list of variable names}

\item{sw_sort_prop}{TRUE/FALSE to sort the last variable descending by proportion within the other variables}
}
\value{
tab_summary a summary table
}
\description{
Create a frequency/proportion summary table by a set of variables
}
\examples{
e_table_sum_freq_prop(
    dat           = ggplot2::mpg
  , var_names     = c("drv", "cyl", "class")
  , sw_sort_prop  = TRUE
  ) \%>\%
  print(n = Inf)
}
