% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_text_filename_replace_characters.R
\name{e_text_filename_replace_characters}
\alias{e_text_filename_replace_characters}
\title{Replace special characters in a filename}
\usage{
e_text_filename_replace_characters(
  fn,
  ch_replace = "_",
  ch_pattern = "[\\\\[\\\\]<>()|:;'&#?*\\\\ ]"
)
}
\arguments{
\item{fn}{list of filenames}

\item{ch_replace}{replacement character for special characters}

\item{ch_pattern}{special characters (regex) to replace}
}
\value{
fn_new     list of filenames with special characters replaced by \code{ch_replace}
}
\description{
Replace special characters in a filename
}
\examples{

fn <- c("dat.txt", "dat[]<>()|:;'&#?* .txt")
fn |> e_text_filename_replace_characters()

}
