% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_text_unicode_to_ascii.R
\name{e_text_unicode_to_ascii}
\alias{e_text_unicode_to_ascii}
\title{Replace Unicode with ASCII}
\usage{
e_text_unicode_to_ascii(
  fn_in = NULL,
  fn_out = NULL,
  unicode_generic_replacement = "XunicodeX"
)
}
\arguments{
\item{fn_in}{text filename in.}

\item{fn_out}{text filename out.  If NULL, then this is assigned \code{fn_in}.}

\item{unicode_generic_replacement}{a string to replace non-matched unicode characters}
}
\value{
NULL, invisibly
}
\description{
Read a text file and replace unicode characters \code{[^\\x00-\\x7F]} with equivalent ASCII or specified replacement.
Review table \code{?dat_unicode_tables_sel_byhand_e} for the ASCII replacements.
Not all unicode has a natural ASCII replacement, so those replacements can be specified with \code{unicode_generic_replacement}.
}
\examples{
\dontrun{
# e_text_unicode_to_ascii(
#     fn_in  = "./data-raw/text_in_unicode.csv"
#   , fn_out = "./data-raw/text_out_unicode.csv"
#   , unicode_generic_replacement = "XxXunicodeXxX"
#   )
}
}
