% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_truncate_limits.R
\name{e_truncate_limits}
\alias{e_truncate_limits}
\title{Truncate limits of a vector to specified min/max values}
\usage{
e_truncate_limits(dat_vec = NULL, min_max = c(-Inf, Inf))
}
\arguments{
\item{dat_vec}{numeric vector}

\item{min_max}{min and max values which to truncate values}
}
\value{
numeric vector
}
\description{
Truncate limits of a vector to specified min/max values
}
\examples{
dat_vec = c(NA, -Inf, -10:10, +Inf, NA)
e_truncate_limits(dat_vec, min_max = c(-Inf, Inf) )
e_truncate_limits(dat_vec, min_max = c(-5, +5)    )
e_truncate_limits(dat_vec, min_max = c(-5, Inf)   )
e_truncate_limits(dat_vec, min_max = c(-Inf, +5)  )
}
