% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ts_bootstrap_phase.R
\name{e_ts_bootstrap_phase}
\alias{e_ts_bootstrap_phase}
\title{Time series bootstrap scramble phase}
\usage{
e_ts_bootstrap_phase(x = NULL)
}
\arguments{
\item{x}{time series object}
}
\value{
x        time series, bootstrapped
}
\description{
Time series bootstrap scramble phase
}
\examples{
e_ts_bootstrap_phase(
   x = datasets::AirPassengers
 )

}
