% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ts_lowpass_filter_ifft_harmonics.R
\name{e_ts_lowpass_filter_ifft_harmonics}
\alias{e_ts_lowpass_filter_ifft_harmonics}
\title{Time series lowpass filter using iFFT harmonics}
\usage{
e_ts_lowpass_filter_ifft_harmonics(
  x = NULL,
  n_harmonics = NULL,
  upsample_multiple = 1
)
}
\arguments{
\item{x}{List of numbers}

\item{n_harmonics}{Number of harmonics}

\item{upsample_multiple}{To upscale number of points.}
}
\value{
x                If \code{upsample_multiple = 1}, then a list of numbers.  If \code{upsample_multiple > 1}, then a dataframe with time points.
}
\description{
Time series lowpass filter using iFFT harmonics
}
\examples{
dat <- log(AirPassengers) |> as.numeric()
dat2 <- e_ts_lowpass_filter_ifft_harmonics(x = dat, n_harmonics = 12)
dat3 <- e_ts_lowpass_filter_ifft_harmonics(x = dat, n_harmonics = 24)
plot(dat |> scale(), type = "l")
points(dat2 |> scale(), type = "l", col = "red")
points(dat3 |> scale(), type = "l", col = "blue")

}
