% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_ttest_pval.R
\name{e_ttest_alt_hyp_text}
\alias{e_ttest_alt_hyp_text}
\title{e_ttest_alt_hyp_text is modified code from stats::print.htest to return text of alternative hypothesis}
\usage{
e_ttest_alt_hyp_text(x, digits = getOption("digits"), prefix = "\\t", ...)
}
\arguments{
\item{x}{stats::t.test object}

\item{...}{}
}
\value{
out text describing alternative hypothesis
}
\description{
e_ttest_alt_hyp_text is modified code from stats::print.htest to return text of alternative hypothesis
}
