% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ttest_plot_pval.R
\name{e_ttest_plot_pval}
\alias{e_ttest_plot_pval}
\title{Plot the result of a t-test by shading the p-value area under the t-distribution}
\usage{
e_ttest_plot_pval(t_summary, sw_graphics = c("ggplot", "base")[2])
}
\arguments{
\item{t_summary}{the returned object from stats::t.test()}

\item{sw_graphics}{choice of ggplot or base graphics}
}
\value{
if base graphics, NULL invisibly; if ggplot, the plot grob
}
\description{
Plot the result of a t-test by shading the p-value area under the t-distribution
}
\examples{
t_summary <- t.test(datasets::mtcars$mpg, mu = 20, data = datasets::mtcars)
e_ttest_plot_pval(t_summary)
t_summary <- t.test(mpg ~ am, mu = 0, data = datasets::mtcars)
e_ttest_plot_pval(t_summary)
}
