% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_vec_winsorize.R
\name{e_vec_winsorize}
\alias{e_vec_winsorize}
\title{Winsorize a numeric list or vector by replacing extreme values with the less extreme values}
\usage{
e_vec_winsorize(x, fraction = 0.05, type = 1)
}
\arguments{
\item{x}{numeric list or vector}

\item{fraction}{fraction between 0 and 0.5 from both sides to Winsorize}

\item{type}{type from \code{quantile()}}
}
\value{
same numeric vector with extreme values replaced
}
\description{
Winsorize a numeric list or vector by replacing extreme values with the less extreme values
}
\examples{

x <- 1:100
y <- e_vec_winsorize(x, fraction = 0.25)
plot(x, y)
}
