% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikmisc.R
\docType{package}
\name{erikmisc-package}
\alias{erikmisc}
\alias{erikmisc-package}
\title{erikmisc: Erik Erhardt's miscellaneous functions for solving complex data analysis workflows}
\description{
Solving common complex data analysis workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erikerhardt/erikmisc}
  \item Report bugs at \url{https://github.com/erikerhardt/erikmisc/issues}
}

}
\author{
\strong{Maintainer}: Erik Erhardt \email{erike@statacumen.com} (\href{https://orcid.org/0000-0002-9817-4011}{ORCID})

}
\keyword{internal}
