% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikmisc_logo.R
\name{erikmisc_logo}
\alias{erikmisc_logo}
\title{The erikmisc logo, using ASCII or Unicode characters}
\usage{
erikmisc_logo(sw_unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{sw_unicode}{Whether to use Unicode symbols. Default is `TRUE`
on UTF-8 platforms.}
}
\description{
Use [crayon::strip_style()] to get rid of the colors.
}
\examples{
erikmisc_logo()
}
