% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{erikmisc_packages}
\alias{erikmisc_packages}
\title{List all packages in the erikmisc}
\usage{
erikmisc_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include erikmisc in the list?}
}
\description{
List all packages in the erikmisc
}
\examples{
erikmisc_packages()
}
