% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_SPECIAL_functions_ALL.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Boolean operator "not in"}
\usage{
... \%notin\% NA
}
\value{

}
\description{
Negation of the \code{\%in\%} operator.
}
\examples{
letters \%notin\% c("a", "b", "c")
}
