% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{mtcars_e}
\alias{mtcars_e}
\title{datasets::mtcars, factored and labeled}
\format{
tibble
}
\source{
\code{?datasets::mtcars}
}
\usage{
mtcars_e
}
\description{
Using \code{datasets::mtcars}, then applying labeled factors and labeling variables with \code{labelled::var_label} based on \code{?datasets::mtcars}
}
\keyword{datasets}
