% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_SPECIAL_functions_ALL.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Boolean operator "not in"}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.}

\item{table}{vector or \code{NULL}: the values to be matched against.}
}
\value{
A logical vector, indicating if a non-match was located for each element of \code{x}: thus the values are \code{TRUE} or \code{FALSE} and never \code{NA}.
}
\description{
Negation of the \code{\%in\%} operator, see \code{?match}.
\code{\%notin\%} is the negation of \code{\link{\%in\%}}, which returns a logical vector indicating if there is a non-match or not for its left operand.
}
\examples{
"a" \%notin\% letters
"a" \%notin\% LETTERS
c("a", "A") \%notin\% letters
letters \%notin\% c("a", "b", "c")
}
