% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erikmisc_logo.R
\name{print.erikmisc_logo}
\alias{print.erikmisc_logo}
\title{Print for erikmisc_logo}
\usage{
\method{print}{erikmisc_logo}(x, ...)
}
\arguments{
\item{x}{the \code{erikmisc_logo()}}

\item{...}{additional parameters passed to \code{cat()}}
}
\value{
the logo, invisibly
}
\description{
Print for erikmisc_logo
}
