% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_complete_multiple_keys_ALL.R
\name{e_coalesce_by_column}
\alias{e_coalesce_by_column}
\title{Combine rows with different data values}
\usage{
e_coalesce_by_column(dat)
}
\arguments{
\item{dat}{data.frame keys to combine}
}
\value{
coalesced data.frame
}
\description{
Combine rows with different data values
}
