% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_data_convert_factor_to_numeric.R
\name{e_data_convert_factor_to_numeric}
\alias{e_data_convert_factor_to_numeric}
\title{Convert all factor variables to their numeric factor levels}
\usage{
e_data_convert_factor_to_numeric(dat)
}
\arguments{
\item{dat}{data.frame or tibble}
}
\value{
dat with factor variables converted to numeric
}
\description{
Before saving to csv, default is the factor labels.  This replaces those with the numeric factor levels.
}
\examples{
dat <- datasets::mtcars \%>\% mutate(cyl = cyl \%>\% factor(), vs = vs \%>\% factor())
str(dat)
dat2 <- dat \%>\% e_data_convert_factor_to_numeric()
str(dat2)
# readr::write_csv(x = dat , file = "dat.csv"    )
# readr::write_csv(x = dat2, file = "dat_num.csv")
}
