% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_SPECIAL_functions_ALL.R
\name{e_extract_numbers_from_string}
\alias{e_extract_numbers_from_string}
\title{Extract numbers from a string}
\usage{
e_extract_numbers_from_string(vec)
}
\arguments{
\item{vec}{list of strings}
}
\value{
numeric list
}
\description{
Extract numbers from a string
}
\examples{
# note difference in last two items
vec <- c(NA, 1, "a", "a1a", "@\%^#@0.23asdf", ")(&*.2&*", ")(&*0.2&*")
e_extract_numbers_from_string(vec)
vec <- c(NA, "1a2")
e_extract_numbers_from_string(vec)
}
