% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_mnv_norm_qqplot.R
\name{e_mnv_norm_qqplot}
\alias{e_mnv_norm_qqplot}
\title{Graphical Assessment (QQ-plot) for assessing Multivariate Normality}
\usage{
e_mnv_norm_qqplot(x, name = "")
}
\arguments{
\item{x}{data.frame or matrix of numeric columns}

\item{name}{label for title}
}
\value{

}
\description{
Graphical Assessment (QQ-plot) for assessing Multivariate Normality
}
\examples{
datasets::mtcars \%>\% filter(cyl == 4) \%>\% select(mpg, hp, wt) \%>\% e_mnv_norm_qqplot(name = "cyl = 4")
datasets::mtcars \%>\% filter(cyl == 6) \%>\% select(mpg, hp, wt) \%>\% e_mnv_norm_qqplot(name = "cyl = 6")
datasets::mtcars \%>\% filter(cyl == 8) \%>\% select(mpg, hp, wt) \%>\% e_mnv_norm_qqplot(name = "cyl = 8")
}
