% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_model_all_subsets_formula.R
\name{e_model_all_subsets_formula}
\alias{e_model_all_subsets_formula}
\title{All model subsets}
\usage{
e_model_all_subsets_formula(
  var_formula = NULL,
  x_var_names = NULL,
  y_var_name = NULL,
  max_scope = c("always", "FO", "TWI", "PQ", "SO")[3],
  sw_return = c("formula", "table")[1]
)
}
\arguments{
\item{var_formula}{NOT YET IMPLEMENTED}

\item{x_var_names}{a character list of variable names or list of character lists.
If a list of multiple formulas or character lists is provided,
then each is subject to the \code{max_scope} independently,
or each can have it's own \code{max_scope}.}

\item{y_var_name}{a character string}

\item{max_scope}{one of or a list (corresponding to \code{x_var_names}) of
"always" to always keep this full list (no binary subsets),
"FO" for first-order (main) effects,
"TWI" for two-way interactions,
"PQ" for pure quadratic terms (squares of the "FO"), and
"SO" for all terms in "FO", "TWI", and "PQ".}

\item{sw_return}{return a list of "formula", or a "table" of x-variable combinations}
}
\value{
\preformatted{         a list of functions
}
}
\description{
Return all possible model subsets based on specified model scopes.
}
\examples{
# Two-way interaction model, return formulas
e_model_all_subsets_formula(
    var_formula = NULL
  , x_var_names = c("a", "b", "c")
  , y_var_name  = "z"
  , max_scope   = "TWI"
  , sw_return   = c("formula", "table")[1]
  )
# Two-way interaction model, return table of x-variable combinations
e_model_all_subsets_formula(
    var_formula = NULL
  , x_var_names = c("a", "b", "c")
  , y_var_name  = "z"
  , max_scope   = "TWI"
  , sw_return   = c("formula", "table")[2]
  )

# Two variables always in the model with a second-order model of two others
e_model_all_subsets_formula(
    var_formula = NULL
  , x_var_names = list(c("a", "b"), c("c", "d"))
  , y_var_name  = "z"
  , max_scope   = list("always", "SO")
  , sw_return   = c("formula", "table")[2]
  )
\dontrun{
# Large example showing all options
e_model_all_subsets_formula(
    var_formula = NULL
  , x_var_names = list(letters[1:2], letters[4:5], letters[7:8], letters[10:11], letters[13:14])
  , y_var_name  = "z"
  , max_scope   = list("always", "FO", "TWI", "PQ", "SO")
  , sw_return   = c("formula", "table")[2]
  )
}

}
