% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_model_extract_var_names}
\alias{e_model_extract_var_names}
\title{From model formula, extract y and x variable lists}
\usage{
e_model_extract_var_names(form = NULL)
}
\arguments{
\item{form}{formula of form y ~ x}
}
\value{
out      list with lists of y variable and x variables
}
\description{
From model formula, extract y and x variable lists
}
