% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_pilot_ALL.R
\name{e_pilot_density_altitude}
\alias{e_pilot_density_altitude}
\title{Compute Koch Chart density altitude, takeoff roll, and climb rate}
\usage{
e_pilot_density_altitude(
  pres_alt = 0,
  temp_C = 15,
  dist_takeoff_sea_level = 1000,
  climb_rate_sea_level = 500,
  sw_propeller = c("fixed_pitch", "constant_speed")[2]
)
}
\arguments{
\item{pres_alt}{pressure altitude, inches of Mg}

\item{temp_C}{temperature, degrees celsius}

\item{dist_takeoff_sea_level}{POH distance for takeoff at
standard temperature and pressure (STP)}

\item{climb_rate_sea_level}{POH distance for climb at
standard temperature and pressure (STP)}

\item{sw_propeller}{Select propeller to be "fixed_pitch" or "constant_speed"}
}
\value{
out                   named list including \code{density_altitude},
\code{dist_takeoff_dens_alt}, and
\code{climb_rate_dens_alt}
}
\description{
Variables:
dens_alt = density altitude, feet
pres_alt = pressure altitude, inches of Mg
temp_C   = temperature, degrees celsius
}
\details{
\code{dens_alt} = (145426 * (1 - (((288.16 - \code{pres_alt} * 0.001981) /
288.16)^5.2563 / ((273.16 + \code{temp_C}) / 288.16))^0.235))

\if{html}{\out{<div class="sourceCode">}}\preformatted{Equations and approximations from "Axioms of Flight",
James Embree, Flight Information Publications,
St. Louis MO, 1984. ISBN 0-9601062-7-8
}\if{html}{\out{</div>}}

APPROXIMATION (good to 8000 feet density altitude):
\itemize{
\item Takeoff distance:
\itemize{
\item For fixed pitch prop, increase sea level standard day takeoff distance
15\% for each 1000 foot increase in density altitude.
\item For constant speed prop, increase by 13\%.
}
\item Climb rate
\itemize{
\item For fixed pitch prop, decrease sea level standard day climb rate
7.5\% for each 1000 foot increase in density altitude.
\item For constant speed prop, decrease by 7\%.
}
}

Example takeoff for fixed pitch prop:
\itemize{
\item dist_takeoff_dens_alt = dist_takeoff_sealevel * (1 + ((dens_alt / 1000) * 0.15))
}

(https://groups.google.com/g/rec.aviation.piloting/c/SDlMioBVqaA)
}
\examples{

e_pilot_density_altitude(
    pres_alt               = 0
  , temp_C                 = 15
  , dist_takeoff_sea_level = 1000
  , climb_rate_sea_level   = 500
  , sw_propeller = c("fixed_pitch", "constant_speed")[1]
  )
e_pilot_density_altitude(
    pres_alt               = 7170
  , temp_C                 = 30
  , dist_takeoff_sea_level = 770
  , climb_rate_sea_level   = 900
  , sw_propeller = c("fixed_pitch", "constant_speed")[1]
  )
e_pilot_density_altitude(
    pres_alt               = 7170
  , temp_C                 = 30
  , dist_takeoff_sea_level = 770
  , climb_rate_sea_level   = 900
  , sw_propeller = c("fixed_pitch", "constant_speed")[2]
  )

}
