% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_complete_by_variable_subset.R
\name{e_plot_complete_by_variable_subset}
\alias{e_plot_complete_by_variable_subset}
\title{For missing data, determine which sets of variables result in the most number of complete observations}
\usage{
e_plot_complete_by_variable_subset(
  dat,
  var_list = NULL,
  var_resp = NULL,
  sw_plot_complete = TRUE
)
}
\arguments{
\item{dat}{data data.frame or tibble}

\item{var_list}{list of variables, \code{NULL} for all}

\item{var_resp}{\code{NULL} or one variable name to always be included (filters to keep only observations with this variable)}

\item{sw_plot_complete}{\code{TRUE} for all, or an integer for the subsets with largest \code{n_complete}}
}
\value{
out      a tibble with the \code{n_complete}, \code{n_var}, \code{var_names_print}, and a list of variable names in \code{var_names}
}
\description{
For missing data, determine which sets of variables result in the most number of complete observations
}
\examples{
# Generate missing values
dat_mtcars_miss_e <- dat_mtcars_e
prop_missing <- 0.10
n_missing <-
  sample.int(
    n    = prod(dim(dat_mtcars_miss_e))
  , size = round( prop_missing * prod(dim(dat_mtcars_miss_e)))
  )
ind_missing <- expand.grid(1:dim(dat_mtcars_miss_e)[1], 1:dim(dat_mtcars_miss_e)[2])[n_missing, ]
for (i_row in seq_along(n_missing)) {
  dat_mtcars_miss_e[ind_missing[i_row, 1], ind_missing[i_row, 2] ] <- NA
}

# Plot missing data
dat_mtcars_miss_e |> e_plot_missing()

out <- dat_mtcars_miss_e |> e_plot_complete_by_variable_subset()
# Print table
out |> print(n = Inf, width = Inf)
# Print variable names from first row
out$var_names[1] |> unlist()

# Print top 5 subsets
dat_mtcars_miss_e |> e_plot_complete_by_variable_subset(sw_plot_complete = 5)
}
