% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_corr_ellipse.R
\name{e_plot_corr_ellipse}
\alias{e_plot_corr_ellipse}
\title{Title}
\usage{
e_plot_corr_ellipse(
  corr,
  outline = FALSE,
  col = "grey",
  upper.panel = c("ellipse", "number", "none")[2],
  lower.panel = c("ellipse", "number", "none")[1],
  diag = c("none", "ellipse", "number")[3],
  digits = 2,
  bty = "n",
  axes = FALSE,
  xlab = "",
  ylab = "",
  asp = 1,
  cex.lab = par("cex.lab"),
  cex = 0.75 * par("cex"),
  mar = c(0, 0, 1, 0),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
