% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_lm_y_covar.R
\name{e_plot_lm_y_covar}
\alias{e_plot_lm_y_covar}
\title{Plot the response vs each predictor in a model formula}
\usage{
e_plot_lm_y_covar(form = form, dat = dat)
}
\arguments{
\item{form}{lm formula for a model}

\item{dat}{data to use}
}
\value{
p ggplot facet scatterplot
}
\description{
Plot the response vs each predictor in a model formula
}
\examples{
# Set specific model with some interactions
form_model <-
  mpg ~ cyl + disp + hp + wt + vs + am + cyl:vs + disp:hp + hp:vs

p <-
  e_plot_lm_y_covar(
    form = form_model
  , dat  = erikmisc::dat_mtcars_e
  )
p |> print()
}
