% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_longitudinal.R
\name{e_plot_longitudinal}
\alias{e_plot_longitudinal}
\title{Longitudinal interaction plots}
\usage{
e_plot_longitudinal(
  dat_plot = dat,
  var_x_time = "x",
  var_y_resp = "y",
  var_ID = "ID",
  var_group = NULL,
  label_x_time = NULL,
  label_y_resp = NULL,
  label_group = NULL,
  x_scale_breaks = NULL,
  y_scale_breaks = NULL,
  hist_scale_breaks = NULL,
  sw_group_reverse = FALSE,
  hist_binwidth = NULL,
  hist_align = c("center", "boundary")[1],
  line_type_grand = c("none", "solid", "dashed", "dotted", "dotdash", "longdash",
    "twodash")[6],
  line_type_group = c("none", "solid", "dashed", "dotted", "dotdash", "longdash",
    "twodash")[4]
)
}
\arguments{
\item{dat_plot}{data.frame for plotting}

\item{var_x_time}{x variable, usually a "time" variable}

\item{var_y_resp}{y response variable}

\item{var_ID}{ID variable for each repeated measures subject}

\item{var_group}{If there are groups, this is the group variable.}

\item{label_x_time}{x-axis label}

\item{label_y_resp}{y-axis label}

\item{label_group}{group label}

\item{x_scale_breaks}{breaks for x-axis}

\item{y_scale_breaks}{breaks for y-axis}

\item{hist_scale_breaks}{breaks for histogram x-axis}

\item{sw_group_reverse}{TRUE/FALSE to reverse the order of groups}

\item{hist_binwidth}{histogram binwidth}

\item{hist_align}{align "center" or "boundary"?  If numeric non-negative responses, usually "boundary" is preferred.}

\item{line_type_grand}{Grand mean line type}

\item{line_type_group}{Group mean line type}
}
\value{
plot in ggplot grob format
}
\description{
Longitudinal interaction plots
}
\examples{
e_plot_longitudinal(
    dat_plot         = lme4::sleepstudy
  , var_x_time       = "Days"
  , var_y_resp       = "Reaction"
  , var_ID           = "Subject"
  , var_group        = NULL
  , label_x_time     = "Number of days of sleep deprivation"
  , label_y_resp     = "Average reaction time (ms)"
  #, label_group      = "Program"
  , x_scale_breaks   = seq(0, 9, by = 2)
  , y_scale_breaks   = seq(0, 1000, by = 50)
  , hist_scale_breaks = NULL
  #, sw_group_reverse = TRUE
  , hist_binwidth    = 25
  , hist_align       = c("center", "boundary")[2]
  , line_type_grand  = c("none", "solid", "dashed", "dotted", "dotdash", "longdash", "twodash")[6]
  , line_type_group  = c("none", "solid", "dashed", "dotted", "dotdash", "longdash", "twodash")[4]
  )
}
