% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_missing.R
\name{e_plot_missing}
\alias{e_plot_missing}
\title{Plots missing data in a data.frame, possibly grouped by one variable and sorted by a second.}
\usage{
e_plot_missing(
  dat_plot,
  var_group = NULL,
  sw_group_sort = FALSE,
  var2_sort = NULL,
  sw_title_data_name = TRUE,
  sw_text_pct_miss = FALSE
)
}
\arguments{
\item{dat_plot}{data.frame or tibble}

\item{var_group}{variable name to group by (colors data)}

\item{sw_group_sort}{TRUE/FALSE to sort by grouped variable}

\item{var2_sort}{second variable name to sort by if data is grouped}

\item{sw_title_data_name}{TRUE/FALSE to include data object name in title or text string of title to use}

\item{sw_text_pct_miss}{TRUE/FALSE to include text values of percent missing on bar plot}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{               ggplot grob plot object
}\if{html}{\out{</div>}}
}
\description{
Plots missing data in a data.frame, possibly grouped by one variable and sorted by a second.
}
\examples{
dat_miss = dat_mtcars_e
prop_missing = 0.10
n_missing = sample.int(n = prod(dim(dat_miss)), size = round( prop_missing * prod(dim(dat_miss))))
ind_missing = expand.grid(1:dim(dat_miss)[1], 1:dim(dat_miss)[2])[n_missing, ]
for (i_row in seq_along(n_missing)) {
  dat_miss[ind_missing[i_row,1], ind_missing[i_row,2] ] <- NA
}

e_plot_missing(
    dat_plot       = dat_miss
  , var_group      = "cyl"
  , sw_group_sort  = TRUE
  , var2_sort      = "disp"
  )

e_plot_missing(
    dat_plot       = dat_miss
  , var_group      = "cyl"
  , sw_group_sort  = TRUE
  , var2_sort      = "disp"
  , sw_title_data_name  = "mtcars with random missing values"
  )
}
