% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__durbinWatsonTest}
\alias{e_plot_model_diagnostics_car__durbinWatsonTest}
\title{Model diagnostics, Durbin-Watson Test for Autocorrelated Errors, car::durbinWatsonTest}
\usage{
e_plot_model_diagnostics_car__durbinWatsonTest(fit = NULL)
}
\arguments{
\item{fit}{fit object}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, Durbin-Watson Test for Autocorrelated Errors, car::durbinWatsonTest
}
