% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__invTranPlot}
\alias{e_plot_model_diagnostics_car__invTranPlot}
\title{Model diagnostics, car::invTranPlot}
\usage{
e_plot_model_diagnostics_car__invTranPlot(fit = NULL, dat = NULL)
}
\arguments{
\item{fit}{fit object}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, car::invTranPlot
}
