% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__outlierTest}
\alias{e_plot_model_diagnostics_car__outlierTest}
\title{Model diagnostics, Bonferroni outlier test, car::outlierTest}
\usage{
e_plot_model_diagnostics_car__outlierTest(fit = NULL)
}
\arguments{
\item{fit}{fit object}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, Bonferroni outlier test, car::outlierTest
}
