% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_car__qqPlot}
\alias{e_plot_model_diagnostics_car__qqPlot}
\title{Model diagnostics, Quantile-Comparison Plot, car::qqPlot}
\usage{
e_plot_model_diagnostics_car__qqPlot(
  fit = NULL,
  dat = NULL,
  sw_interp = c(TRUE, FALSE)[2]
)
}
\arguments{
\item{fit}{fit object}

\item{dat}{dataset data.frame or tibble}

\item{sw_interp}{T/F to provide interpretation guidance in the plot caption or with a table.}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, Quantile-Comparison Plot, car::qqPlot
}
