% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_model_diagnostics_EACH.R
\name{e_plot_model_diagnostics_Resid_histogram}
\alias{e_plot_model_diagnostics_Resid_histogram}
\title{Model diagnostics, Cooks vs Leverage vs Residuals}
\usage{
e_plot_model_diagnostics_Resid_histogram(fit_resid = NULL)
}
\arguments{
\item{fit_resid}{list of residuals from \code{e_model_calc_resid()}}
}
\value{
out      list including text and ggplot grobs
}
\description{
Model diagnostics, Cooks vs Leverage vs Residuals
}
