% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_rf_vimp.R
\name{e_plot_rf_vimp}
\alias{e_plot_rf_vimp}
\title{A function to plot the "all" column from a Random Forests Variable Importance Plot (VIMP).}
\usage{
e_plot_rf_vimp(v_imp)
}
\arguments{
\item{v_imp}{\code{$importance} table from \code{randomForestSRC::rfsrc()}}
}
\value{
p
}
\description{
A function to plot the "all" column from a Random Forests Variable Importance Plot (VIMP).
}
\examples{
# o_class <- randomForestSRC::rfsrc(...)
# e_plot_rf_vimp(o_class$importance)
}
