% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_plot_scatterplot.R
\name{e_plot_scatterplot}
\alias{e_plot_scatterplot}
\title{Plot scatterplot of a numeric y-variable against a numeric or categorical x-variable}
\usage{
e_plot_scatterplot(
  dat_plot = NULL,
  var_x = NULL,
  var_y = NULL,
  var_color = NULL,
  var_facet = NULL,
  text_title = NULL,
  sw_print = FALSE,
  smooth_all = c("none", "loess", "lm", "glm", "gam")[1],
  sw_smooth_all_se = c(TRUE, FALSE)[1],
  smooth_by_var_color = c("none", "loess", "lm", "glm", "gam")[1],
  sw_smooth_by_var_color_se = c(TRUE, FALSE)[1],
  sw_corr_in_subtitle = c(TRUE, FALSE)[1]
)
}
\arguments{
\item{dat_plot}{data to plot}

\item{var_x}{a numeric or categorical variable name}

\item{var_y}{a numeric variable name}

\item{var_color}{factor or numeric color variable}

\item{var_facet}{factor varibles (1 or 2) to facet by, (row facets, then column facets),}

\item{sw_print}{T/F whether to print table and display plot}

\item{smooth_all}{Smooth method for all the points together.}

\item{sw_smooth_all_se}{T/F, SEs for smooth all the points together?}

\item{smooth_by_var_color}{Smooth method by the color variable.}

\item{sw_smooth_by_var_color_se}{T/F, SEs for smooth by the color variable?}

\item{sw_corr_in_subtitle}{T/F, calculate correlation and include in subtitle of plot?}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a ggplot object
}\if{html}{\out{</div>}}
}
\description{
Can add up to one color factor or numeric variable and two facets factor variables, and choose to display correlation.
}
\examples{
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "disp"           # numeric x
  , var_y     = "mpg"
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "am"             # factor x
  , var_y     = "mpg"
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "disp"
  , var_y     = "mpg"
  , var_color = "cyl"            # factor color
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "disp"
  , var_y     = "mpg"
  , var_color = "cyl"            # factor color
  , smooth_all = c("none", "loess", "lm", "glm", "gam")[2]  # with all smooth
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "disp"
  , var_y     = "mpg"
  , var_color = "cyl"            # factor color
  , smooth_by_var_color = c("none", "loess", "lm", "glm", "gam")[3]  # with group smooth
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "disp"
  , var_y     = "mpg"
  , var_color = "wt"             # numeric color
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "disp"
  , var_y     = "mpg"
  , var_facet = c("vs")          # one facet
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "disp"
  , var_y     = "mpg"
  , var_color = "cyl"
  , var_facet = c("vs", "am")    # two facets
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "am"             # factor x
  , var_y     = "mpg"
  , var_color = "wt"             # numeric color
  , var_facet = c("vs")
  )
e_plot_scatterplot(
    dat_plot  = dat_mtcars_e
  , var_x     = "am"             # factor x
  , var_y     = "mpg"
  , var_color = "vs"             # factor color
  )
}
