% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_read_data_files.R
\name{e_read_data_files}
\alias{e_read_data_files}
\title{Read a list of csv, xls, or xlsx data into a list of tibbles}
\usage{
e_read_data_files(
  read_fn_path = ".",
  read_fn_names = NULL,
  sw_dat_add_col_path_fn = c(TRUE, FALSE)[1],
  sw_dat_print_fn_read = c(TRUE, FALSE)[1],
  excel_sheets = "all",
  sw_clean_names = c(TRUE, FALSE)[2],
  excel_range = NULL,
  excel_col_names = TRUE
)
}
\arguments{
\item{read_fn_path}{Directory path}

\item{read_fn_names}{Character list of file names}

\item{sw_dat_add_col_path_fn}{T/F for data, add two columns specifying the directory (\code{DIR__}) and filename (\code{FILE__})}

\item{sw_dat_print_fn_read}{T/F print file names and dimensions as the files are read}

\item{excel_sheets}{"all" for all sheets, or a list of numbers "\code{c(1, 2)}"; applies to all excel sheets.}

\item{sw_clean_names}{T/F to clean column names using \code{janitor::clean_names}}

\item{excel_range}{When reading Excel files, NULL reads entire sheet, a range is specified as in \code{readxl::read_xlsx}.  Applies to all files.}

\item{excel_col_names}{Specified as in \code{readxl::read_xlsx}.  Applies to all files.}
}
\value{
dat_sheet             A list of tibbles
}
\description{
Called from \code{e_read_data_subdir_into_lists()}.
}
\examples{
\dontrun{
# See e_read_data_subdir_into_lists()
}
}
