% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_REDCap_read_format_data.R
\name{e_REDCap_read_format_data}
\alias{e_REDCap_read_format_data}
\title{REDCap data, reading and formatting data}
\usage{
e_REDCap_read_format_data(
  fn_path = getwd(),
  fn_root = NULL,
  fn_suffix_data = "_DATA_",
  fn_suffix_R = "_R_",
  fn_suffix_out = "",
  fn_ext_data = "csv",
  fn_ext_R = "r",
  sw_return_dat = TRUE,
  sw_save_RData = TRUE
)
}
\arguments{
\item{fn_path}{is the location of the data, assumed to be the current path.}

\item{fn_root}{the root name(s) of the file(s), typically the project name(s) (the part before \emph{DATA}\emph{.})}

\item{fn_suffix_data}{typically "\emph{DATA}" to indicate this is the data}

\item{fn_suffix_R}{typically "\emph{R}" to indicate this is the R formatting file}

\item{fn_suffix_out}{an added name to the outputted saved and formatted data files, maybe you want to add a date, a version number, or a another label}

\item{fn_ext_data}{"*.csv" to specify the data filename extension, only csv is supported ("csv" is differnt from "CSV")}

\item{fn_ext_R}{"*.r" to specify the R filename extension, ("r" is differnt from "R")}

\item{sw_return_dat}{TRUE/FALSE to return the data object from the function}

\item{sw_save_RData}{TRUE/FALSE to save an ".RData" file}
}
\value{
out_dat formatted data or \code{invisible(NULL)}, depending on \code{sw_return_dat}
}
\description{
Read the files in the fn_path and find the most recent file (last
alphabetically, which is by date based on REDCap file name convention of
\emph{_yyyy-mm-dd_hhmm.}) and read and format that given the R formatting file.
}
\details{
For each project file:
\itemize{
\item Menu > Reports (Data Exports, Reports, and Stats) > All data
\item All data (all records and fields), Export Data
\item R Statistical Software
\item Click both of the R and CSV file icons
\itemize{
\item a. Save to your path
\item b. Name as: (default name) \link{fn_root}_DATA_yyyy-mm-dd_hhmm.csv, \link{fn_root}_R_yyyy-mm-dd_hhmm.csv
}
}
}
\examples{
\dontrun{
# dat_temp <-
#   e_REDCap_read_format_data(
#     fn_path   = "C:/Data/Project"
#   , fn_root   = "ProjectName"
#   )
}
}
