% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_table_print.R
\name{e_table_print}
\alias{e_table_print}
\title{Print a data.frame table for latex and scale to fit page}
\usage{
e_table_print(dat, sw_scale = FALSE, sw_kable_format = c("html", "latex")[1])
}
\arguments{
\item{dat}{data.frame or tibble to print}

\item{sw_scale}{if "latex" T/F to scale to fit page, if "html" then a point size to pass to \code{kableExtra::kable_styling} argument \code{font_size}}

\item{sw_kable_format}{"html" or "latex" format}
}
\value{

}
\description{
If using in Rmd file and a latex documnet, then requires two header includes:
}
\details{
\preformatted{
EXAMPLE OF .Rmd YAML header

---
title: "Title"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  pdf_document:
    df_print: kable
    latex_engine: lualatex
    toc: true
    number_sections: true
    toc_depth: 2
    keep_tex: true
header-includes:
- \usepackage{booktabs}
- \usepackage{colortbl}
always_allow_html: yes
fig_caption: TRUE
---
}
}
\examples{
datasets::mtcars \%>\% head() \%>\% e_table_print()
\dontrun{
datasets::mtcars \%>\% head() \%>\% e_table_print(sw_scale = 6)
datasets::mtcars \%>\% head() \%>\% e_table_print(sw_scale = TRUE, sw_kable_format = "latex") # outputs into LaTeX document
}
}
