% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ttest_plot_pval.R
\name{e_ttest_alt_hyp_text}
\alias{e_ttest_alt_hyp_text}
\title{Return text of alternative hypothesis via modified code from stats::print.htest}
\usage{
e_ttest_alt_hyp_text(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{stats::t.test object}

\item{digits}{number of digits to return, from \code{getOption("digits")}}

\item{...}{additional parameters}
}
\value{
out text describing alternative hypothesis
}
\description{
Return text of alternative hypothesis via modified code from stats::print.htest
}
