% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_SPECIAL_functions_ALL.R
\name{notin}
\alias{notin}
\alias{\%notin\%}
\title{Boolean operator "not in"}
\usage{
x \%notin\% y

notin(x, y)
}
\arguments{
\item{x}{Vector or NULL: the values to be matched.}

\item{y}{Vector or NULL: the values to be matched against.}
}
\value{
A logical vector of the same length as \code{x}.

  For every element in \code{x}, return \code{FALSE} if the value exists
  in \code{y}, and \code{TRUE} otherwise.
}
\description{
Negation of the \code{\%in\%} operator.
Determine if values in the first argument don't exist in the second argument.
}
\details{
\code{\%notin\%} is the functional inverse of \code{\%in\%}.

\code{lhs \%notin\% rhs} is equivalent to \code{notin(lhs, rhs)}.
}
\examples{
"a" \%notin\% letters
"a" \%notin\% LETTERS
c("a", "A") \%notin\% letters
letters \%notin\% c("a", "b", "c")
notin("A", letters)
}
